@@ -59,17 +59,16 @@ use Symfony\Component\HttpFoundation\StreamedResponse;
 use Symfony\Component\HttpKernel\Event\ControllerEvent;
 use Symfony\Component\Mime\MimeTypes;
 use Symfony\Component\Process\Process;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Twig\Environment;
 use Twig\Extension\CoreExtension;
 
 /**
- * @Route("/asset")
- *
  * @internal
  */
+#[Route("/asset")]
 class AssetController extends ElementControllerBase implements KernelControllerEventInterface
 {
     use AdminStyleTrait;
@@ -81,25 +80,19 @@ class AssetController extends ElementControllerBase implements KernelControllerE
 
     protected Asset\Service $_assetService;
 
-    /**
-     * @Route("/tree-get-root", name="opendxp_admin_asset_treegetroot", methods={"GET"})
-     */
+    #[Route("/tree-get-root", name: "opendxp_admin_asset_treegetroot", methods: ["GET"])]
     public function treeGetRootAction(Request $request): JsonResponse
     {
         return parent::treeGetRootAction($request);
     }
 
-    /**
-     * @Route("/delete-info", name="opendxp_admin_asset_deleteinfo", methods={"GET"})
-     */
+    #[Route("/delete-info", name: "opendxp_admin_asset_deleteinfo", methods: ["GET"])]
     public function deleteInfoAction(Request $request, EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         return parent::deleteInfoAction($request, $eventDispatcher);
     }
 
-    /**
-     * @Route("/get-data-by-id", name="opendxp_admin_asset_getdatabyid", methods={"GET"})
-     */
+    #[Route("/get-data-by-id", name: "opendxp_admin_asset_getdatabyid", methods: ["GET"])]
     public function getDataByIdAction(Request $request, EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         $assetId = $request->query->getInt('id');
@@ -246,9 +239,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         throw $this->createAccessDeniedHttpException();
     }
 
-    /**
-     * @Route("/tree-get-children-by-id", name="opendxp_admin_asset_treegetchildrenbyid", methods={"GET"})
-     */
+    #[Route("/tree-get-children-by-id", name: "opendxp_admin_asset_treegetchildrenbyid", methods: ["GET"])]
     public function treeGetChildrenByIdAction(Request $request, EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         $allParams = array_merge($request->request->all(), $request->query->all());
@@ -337,9 +328,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         }
     }
 
-    /**
-     * @Route("/add-asset", name="opendxp_admin_asset_addasset", methods={"POST"})
-     */
+    #[Route("/add-asset", name: "opendxp_admin_asset_addasset", methods: ["POST"])]
     public function addAssetAction(Request $request, Config $config): JsonResponse
     {
         try {
@@ -366,9 +355,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         }
     }
 
-    /**
-     * @Route("/add-asset-compatibility", name="opendxp_admin_asset_addassetcompatibility", methods={"POST"})
-     */
+    #[Route("/add-asset-compatibility", name: "opendxp_admin_asset_addassetcompatibility", methods: ["POST"])]
     public function addAssetCompatibilityAction(Request $request, Config $config): JsonResponse
     {
         try {
@@ -394,10 +381,9 @@ class AssetController extends ElementControllerBase implements KernelControllerE
     }
 
     /**
-     * @Route("/exists", name="opendxp_admin_asset_exists", methods={"GET"})
-     *
      * @throws \Exception
      */
+    #[Route("/exists", name: "opendxp_admin_asset_exists", methods: ["GET"])]
     public function existsAction(Request $request): JsonResponse
     {
         $parentAsset = \OpenDxp\Model\Asset::getById((int)$request->get('parentId'));
@@ -581,10 +567,9 @@ class AssetController extends ElementControllerBase implements KernelControllerE
     }
 
     /**
-     * @Route("/replace-asset", name="opendxp_admin_asset_replaceasset", methods={"POST", "PUT"})
-     *
      * @throws \Exception
      */
+    #[Route("/replace-asset", name: "opendxp_admin_asset_replaceasset", methods: ["POST", "PUT"])]
     public function replaceAssetAction(Request $request, TranslatorInterface $translator): JsonResponse
     {
         $asset = Asset::getById((int) $request->get('id'));
@@ -635,9 +620,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         }
     }
 
-    /**
-     * @Route("/add-folder", name="opendxp_admin_asset_addfolder", methods={"POST"})
-     */
+    #[Route("/add-folder", name: "opendxp_admin_asset_addfolder", methods: ["POST"])]
     public function addFolderAction(Request $request): JsonResponse
     {
         $success = false;
@@ -661,9 +644,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         return $this->adminJson(['success' => $success]);
     }
 
-    /**
-     * @Route("/delete", name="opendxp_admin_asset_delete", methods={"DELETE"})
-     */
+    #[Route("/delete", name: "opendxp_admin_asset_delete", methods: ["DELETE"])]
     public function deleteAction(Request $request): JsonResponse
     {
         $type = $request->get('type');
@@ -715,11 +696,10 @@ class AssetController extends ElementControllerBase implements KernelControllerE
     }
 
     /**
-     * @Route("/update", name="opendxp_admin_asset_update", methods={"PUT"})
-     *
      * @throws \Exception
      * @throws RuntimeException
      */
+    #[Route("/update", name: "opendxp_admin_asset_update", methods: ["PUT"])]
     public function updateAction(Request $request): JsonResponse
     {
         $data = ['success' => false];
@@ -803,10 +783,9 @@ class AssetController extends ElementControllerBase implements KernelControllerE
     }
 
     /**
-     * @Route("/save", name="opendxp_admin_asset_save", methods={"PUT","POST"})
-     *
      * @throws \Exception
      */
+    #[Route("/save", name: "opendxp_admin_asset_save", methods: ["PUT","POST"])]
     public function saveAction(Request $request, EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         $asset = Asset::getById((int) $request->get('id'));
@@ -904,9 +883,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         }
     }
 
-    /**
-     * @Route("/publish-version", name="opendxp_admin_asset_publishversion", methods={"POST"})
-     */
+    #[Route("/publish-version", name: "opendxp_admin_asset_publishversion", methods: ["POST"])]
     public function publishVersionAction(Request $request): JsonResponse
     {
         $id = (int)$request->get('id');
@@ -934,9 +911,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         throw $this->createAccessDeniedHttpException();
     }
 
-    /**
-     * @Route("/show-version", name="opendxp_admin_asset_showversion", methods={"GET"})
-     */
+    #[Route("/show-version", name: "opendxp_admin_asset_showversion", methods: ["GET"])]
     public function showVersionAction(Request $request, Environment $twig): Response
     {
         $id = (int)$request->get('id');
@@ -978,9 +953,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         );
     }
 
-    /**
-     * @Route("/download", name="opendxp_admin_asset_download", methods={"GET"})
-     */
+    #[Route("/download", name: "opendxp_admin_asset_download", methods: ["GET"])]
     public function downloadAction(Request $request): StreamedResponse
     {
         $asset = Asset::getById((int) $request->get('id'));
@@ -1003,14 +976,12 @@ class AssetController extends ElementControllerBase implements KernelControllerE
             fpassthru($stream);
         }, 200, [
             'Content-Type' => $asset->getMimeType(),
-            'Content-Disposition' => sprintf('attachment; filename="%s"', $asset->getFilename()),
+            'Content-Disposition' => sprintf('attachment; filename: "%s"', $asset->getFilename()),
             'Content-Length' => $asset->getFileSize(),
         ]);
     }
 
-    /**
-     * @Route("/download-image-thumbnail", name="opendxp_admin_asset_downloadimagethumbnail", methods={"GET"})
-     */
+    #[Route("/download-image-thumbnail", name: "opendxp_admin_asset_downloadimagethumbnail", methods: ["GET"])]
     public function downloadImageThumbnailAction(Request $request): BinaryFileResponse
     {
         $image = Asset\Image::getById((int) $request->get('id'));
@@ -1139,9 +1110,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         throw $this->createNotFoundException('Thumbnail not found');
     }
 
-    /**
-     * @Route("/get-asset", name="opendxp_admin_asset_getasset", methods={"GET"})
-     */
+    #[Route("/get-asset", name: "opendxp_admin_asset_getasset", methods: ["GET"])]
     public function getAssetAction(Request $request): StreamedResponse
     {
         $image = Asset::getById((int)$request->get('id'));
@@ -1171,9 +1140,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         return $response;
     }
 
-    /**
-     * @Route("/get-image-thumbnail", name="opendxp_admin_asset_getimagethumbnail", methods={"GET"})
-     */
+    #[Route("/get-image-thumbnail", name: "opendxp_admin_asset_getimagethumbnail", methods: ["GET"])]
     public function getImageThumbnailAction(Request $request): BinaryFileResponse|JsonResponse|StreamedResponse
     {
         $fileinfo = $request->get('fileinfo');
@@ -1260,9 +1227,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         return $response;
     }
 
-    /**
-     * @Route("/get-folder-thumbnail", name="opendxp_admin_asset_getfolderthumbnail", methods={"GET"})
-     */
+    #[Route("/get-folder-thumbnail", name: "opendxp_admin_asset_getfolderthumbnail", methods: ["GET"])]
     public function getFolderThumbnailAction(Request $request): StreamedResponse
     {
         $folder = null;
@@ -1294,9 +1259,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         throw $this->createNotFoundException('could not load asset folder');
     }
 
-    /**
-     * @Route("/get-video-thumbnail", name="opendxp_admin_asset_getvideothumbnail", methods={"GET"})
-     */
+    #[Route("/get-video-thumbnail", name: "opendxp_admin_asset_getvideothumbnail", methods: ["GET"])]
     public function getVideoThumbnailAction(Request $request): StreamedResponse
     {
         $video = null;
@@ -1369,9 +1332,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         return $response;
     }
 
-    /**
-     * @Route("/get-document-thumbnail", name="opendxp_admin_asset_getdocumentthumbnail", methods={"GET"})
-     */
+    #[Route("/get-document-thumbnail", name: "opendxp_admin_asset_getdocumentthumbnail", methods: ["GET"])]
     public function getDocumentThumbnailAction(Request $request): BinaryFileResponse|StreamedResponse
     {
         $document = Asset\Document::getById((int)$request->get('id'));
@@ -1438,9 +1399,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         $response->headers->set('Pragma', '');
     }
 
-    /**
-     * @Route("/get-preview-document", name="opendxp_admin_asset_getpreviewdocument", methods={"GET"})
-     */
+    #[Route("/get-preview-document", name: "opendxp_admin_asset_getpreviewdocument", methods: ["GET"])]
     public function getPreviewDocumentAction(
         Request $request,
         TranslatorInterface $translator
@@ -1537,9 +1496,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         return $stream;
     }
 
-    /**
-     * @Route("/get-preview-video", name="opendxp_admin_asset_getpreviewvideo", methods={"GET"})
-     */
+    #[Route("/get-preview-video", name: "opendxp_admin_asset_getpreviewvideo", methods: ["GET"])]
     public function getPreviewVideoAction(Request $request): Response
     {
         $asset = Asset\Video::getById((int) $request->get('id'));
@@ -1587,9 +1544,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         }
     }
 
-    /**
-     * @Route("/serve-video-preview", name="opendxp_admin_asset_servevideopreview", methods={"GET"})
-     */
+    #[Route("/serve-video-preview", name: "opendxp_admin_asset_servevideopreview", methods: ["GET"])]
     public function serveVideoPreviewAction(Request $request): StreamedResponse
     {
         $asset = Asset\Video::getById((int) $request->get('id'));
@@ -1629,9 +1584,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         }
     }
 
-    /**
-     * @Route("/image-editor", name="opendxp_admin_asset_imageeditor", methods={"GET"})
-     */
+    #[Route("/image-editor", name: "opendxp_admin_asset_imageeditor", methods: ["GET"])]
     public function imageEditorAction(Request $request): Response
     {
         $asset = Asset::getById((int) $request->get('id'));
@@ -1646,9 +1599,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         );
     }
 
-    /**
-     * @Route("/image-editor-save", name="opendxp_admin_asset_imageeditorsave", methods={"PUT"})
-     */
+    #[Route("/image-editor-save", name: "opendxp_admin_asset_imageeditorsave", methods: ["PUT"])]
     public function imageEditorSaveAction(Request $request): JsonResponse
     {
         $asset = Asset::getById((int) $request->get('id'));
@@ -1671,12 +1622,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/get-folder-content-preview", name="opendxp_admin_asset_getfoldercontentpreview", methods={"GET"})
-     */
-    /**
-     * @Route("/get-folder-content-preview", name="opendxp_admin_asset_getfoldercontentpreview", methods={"GET"})
-     */
+    #[Route("/get-folder-content-preview", name: "opendxp_admin_asset_getfoldercontentpreview", methods: ["GET"])]
     public function getFolderContentPreviewAction(Request $request,
         EventDispatcherInterface $eventDispatcher,
         GridHelperService $gridHelperService): JsonResponse
@@ -1764,9 +1710,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         return $this->adminJson(['assets' => $result['data'], 'success' => $result['success'], 'total' => $result['total']]);
     }
 
-    /**
-     * @Route("/copy-info", name="opendxp_admin_asset_copyinfo", methods={"GET"})
-     */
+    #[Route("/copy-info", name: "opendxp_admin_asset_copyinfo", methods: ["GET"])]
     public function copyInfoAction(Request $request): JsonResponse
     {
         $transactionId = time();
@@ -1839,9 +1783,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         ]);
     }
 
-    /**
-     * @Route("/copy", name="opendxp_admin_asset_copy", methods={"POST"})
-     */
+    #[Route("/copy", name: "opendxp_admin_asset_copy", methods: ["POST"])]
     public function copyAction(Request $request): JsonResponse
     {
         $success = false;
@@ -1857,7 +1799,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
 
             // this is because the key can get the prefix "_copy" if the target does already exists
             if ($sessionBag['parentId']) {
-                $targetParent = Asset::getById($sessionBag['parentId']);
+                $targetParent = Asset::getById((int) $sessionBag['parentId']);
             } else {
                 $targetParent = Asset::getById((int) $request->get('targetParentId'));
             }
@@ -1901,9 +1843,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         return $this->adminJson(['success' => $success]);
     }
 
-    /**
-     * @Route("/download-as-zip-jobs", name="opendxp_admin_asset_downloadaszipjobs", methods={"GET"})
-     */
+    #[Route("/download-as-zip-jobs", name: "opendxp_admin_asset_downloadaszipjobs", methods: ["GET"])]
     public function downloadAsZipJobsAction(Request $request): JsonResponse
     {
         $jobId = uniqid();
@@ -1974,9 +1914,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         ]);
     }
 
-    /**
-     * @Route("/download-as-zip-add-files", name="opendxp_admin_asset_downloadaszipaddfiles", methods={"GET"})
-     */
+    #[Route("/download-as-zip-add-files", name: "opendxp_admin_asset_downloadaszipaddfiles", methods: ["GET"])]
     public function downloadAsZipAddFilesAction(Request $request): JsonResponse
     {
         $zipFile = OPENDXP_SYSTEM_TEMP_DIRECTORY . '/download-zip-' . $request->get('jobId') . '.zip';
@@ -2058,9 +1996,8 @@ class AssetController extends ElementControllerBase implements KernelControllerE
     /**
      * Download all assets contained in the folder with parameter id as ZIP file.
      * The suggested filename is either [folder name].zip or assets.zip for the root folder.
-     *
-     * @Route("/download-as-zip", name="opendxp_admin_asset_downloadaszip", methods={"GET"})
      */
+    #[Route("/download-as-zip", name: "opendxp_admin_asset_downloadaszip", methods: ["GET"])]
     public function downloadAsZipAction(Request $request): BinaryFileResponse
     {
         $asset = Asset::getById((int) $request->get('id'));
@@ -2081,9 +2018,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         return $response;
     }
 
-    /**
-     * @Route("/import-zip", name="opendxp_admin_asset_importzip", methods={"POST"})
-     */
+    #[Route("/import-zip", name: "opendxp_admin_asset_importzip", methods: ["POST"])]
     public function importZipAction(Request $request, TranslatorInterface $translator): Response
     {
         $jobId = uniqid();
@@ -2148,9 +2083,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         }
     }
 
-    /**
-     * @Route("/import-zip-files", name="opendxp_admin_asset_importzipfiles", methods={"POST"})
-     */
+    #[Route("/import-zip-files", name: "opendxp_admin_asset_importzipfiles", methods: ["POST"])]
     public function importZipFilesAction(Request $request, Filesystem $filesystem): JsonResponse
     {
         $jobId = $request->get('jobId');
@@ -2226,9 +2159,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         ]);
     }
 
-    /**
-     * @Route("/clear-thumbnail", name="opendxp_admin_asset_clearthumbnail", methods={"POST"})
-     */
+    #[Route("/clear-thumbnail", name: "opendxp_admin_asset_clearthumbnail", methods: ["POST"])]
     public function clearThumbnailAction(Request $request): JsonResponse
     {
         $success = false;
@@ -2249,9 +2180,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         return $this->adminJson(['success' => $success]);
     }
 
-    /**
-     * @Route("/grid-proxy", name="opendxp_admin_asset_gridproxy", methods={"GET", "POST", "PUT"})
-     */
+    #[Route("/grid-proxy", name: "opendxp_admin_asset_gridproxy", methods: ["GET", "POST", "PUT"])]
     public function gridProxyAction(Request $request, EventDispatcherInterface $eventDispatcher, GridHelperService $gridHelperService, CsrfProtectionHandler $csrfProtection): JsonResponse
     {
         $allParams = array_merge($request->request->all(), $request->query->all());
@@ -2290,7 +2219,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
                     $data = $updateEvent->getArgument('data');
 
                     // save
-                    $asset = Asset::getById($data['id']);
+                    $asset = Asset::getById((int) $data['id']);
 
                     if (!$asset) {
                         throw $this->createNotFoundException('Asset not found');
@@ -2428,9 +2357,7 @@ class AssetController extends ElementControllerBase implements KernelControllerE
         return $this->adminJson(['success' => false]);
     }
 
-    /**
-     * @Route("/get-text", name="opendxp_admin_asset_gettext", methods={"GET"})
-     */
+    #[Route("/get-text", name: "opendxp_admin_asset_gettext", methods: ["GET"])]
     public function getTextAction(Request $request): JsonResponse
     {
         $asset = Asset::getById((int) $request->get('id'));
