@@ -44,14 +44,14 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Session\Attribute\AttributeBagInterface;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
 /**
- * @Route("/asset-helper")
- *
  * @internal
  */
+
+#[Route("/asset-helper")]
 class AssetHelperController extends AdminAbstractController
 {
     public function __construct(
@@ -123,9 +123,7 @@ class AssetHelperController extends AdminAbstractController
         return $configData;
     }
 
-    /**
-     * @Route("/grid-delete-column-config", name="opendxp_admin_asset_assethelper_griddeletecolumnconfig", methods={"DELETE"})
-     */
+    #[Route("/grid-delete-column-config", name: "opendxp_admin_asset_assethelper_griddeletecolumnconfig", methods: ["DELETE"])]
     public function gridDeleteColumnConfigAction(Request $request): JsonResponse
     {
         $gridConfigId = (int) $request->get('gridConfigId');
@@ -146,9 +144,7 @@ class AssetHelperController extends AdminAbstractController
         return $this->adminJson($newGridConfig);
     }
 
-    /**
-     * @Route("/grid-get-column-config", name="opendxp_admin_asset_assethelper_gridgetcolumnconfig", methods={"GET"})
-     */
+    #[Route("/grid-get-column-config", name: "opendxp_admin_asset_assethelper_gridgetcolumnconfig", methods: ["GET"])]
     public function gridGetColumnConfigAction(Request $request): JsonResponse
     {
         $result = $this->doGetGridColumnConfig($request);
@@ -355,9 +351,7 @@ class AssetHelperController extends AdminAbstractController
         return $availableFields;
     }
 
-    /**
-     * @Route("/prepare-helper-column-configs", name="opendxp_admin_asset_assethelper_preparehelpercolumnconfigs", methods={"POST"})
-     */
+    #[Route("/prepare-helper-column-configs", name: "opendxp_admin_asset_assethelper_preparehelpercolumnconfigs", methods: ["POST"])]
     public function prepareHelperColumnConfigs(Request $request): JsonResponse
     {
         $helperColumns = [];
@@ -385,13 +379,11 @@ class AssetHelperController extends AdminAbstractController
         return $this->adminJson(['success' => true, 'columns' => $newData]);
     }
 
-    /**
-     * @Route("/grid-mark-favourite-column-config", name="opendxp_admin_asset_assethelper_gridmarkfavouritecolumnconfig", methods={"POST"})
-     */
+    #[Route("/grid-mark-favourite-column-config", name: "opendxp_admin_asset_assethelper_gridmarkfavouritecolumnconfig", methods: ["POST"])]
     public function gridMarkFavouriteColumnConfigAction(Request $request): JsonResponse
     {
         $classId = $request->get('classId');
-        $asset = Asset::getById($classId);
+        $asset = Asset::getById(is_numeric($classId) ? (int) $classId : 0);
 
         if ($asset->isAllowed('list')) {
             $gridConfigId = (int) $request->get('gridConfigId');
@@ -451,9 +443,7 @@ class AssetHelperController extends AdminAbstractController
         return $result;
     }
 
-    /**
-     * @Route("/grid-save-column-config", name="opendxp_admin_asset_assethelper_gridsavecolumnconfig", methods={"POST"})
-     */
+    #[Route("/grid-save-column-config", name: "opendxp_admin_asset_assethelper_gridsavecolumnconfig", methods: ["POST"])]
     public function gridSaveColumnConfigAction(Request $request): JsonResponse
     {
         $asset = Asset::getById((int) $request->get('id'));
@@ -653,9 +643,7 @@ class AssetHelperController extends AdminAbstractController
         }
     }
 
-    /**
-     * @Route("/get-export-jobs", name="opendxp_admin_asset_assethelper_getexportjobs", methods={"POST"})
-     */
+    #[Route("/get-export-jobs", name: "opendxp_admin_asset_assethelper_getexportjobs", methods: ["POST"])]
     public function getExportJobsAction(Request $request, GridHelperService $gridHelperService): JsonResponse
     {
         $allParams = array_merge($request->request->all(), $request->query->all());
@@ -675,10 +663,10 @@ class AssetHelperController extends AdminAbstractController
     }
 
     /**
-     * @Route("/do-export", name="opendxp_admin_asset_assethelper_doexport", methods={"POST"})
-     *
      * @throws FilesystemException
      */
+
+    #[Route("/do-export", name: "opendxp_admin_asset_assethelper_doexport", methods: ["POST"])]
     public function doExportAction(Request $request): JsonResponse
     {
         $fileHandle = File::getValidFilename($request->get('fileHandle'));
@@ -832,9 +820,7 @@ class AssetHelperController extends AdminAbstractController
         return $fileHandle . '.csv';
     }
 
-    /**
-     * @Route("/download-csv-file", name="opendxp_admin_asset_assethelper_downloadcsvfile", methods={"GET"})
-     */
+    #[Route("/download-csv-file", name: "opendxp_admin_asset_assethelper_downloadcsvfile", methods: ["GET"])]
     public function downloadCsvFileAction(Request $request): Response
     {
         $storage = Storage::get('temp');
@@ -860,9 +846,7 @@ class AssetHelperController extends AdminAbstractController
         }
     }
 
-    /**
-     * @Route("/download-xlsx-file", name="opendxp_admin_asset_assethelper_downloadxlsxfile", methods={"GET"})
-     */
+    #[Route("/download-xlsx-file", name: "opendxp_admin_asset_assethelper_downloadxlsxfile", methods: ["GET"])]
     public function downloadXlsxFileAction(Request $request, GridHelperService $gridHelperService): BinaryFileResponse
     {
         $storage = Storage::get('temp');
@@ -877,9 +861,7 @@ class AssetHelperController extends AdminAbstractController
         }
     }
 
-    /**
-     * @Route("/get-metadata-for-column-config", name="opendxp_admin_asset_assethelper_getmetadataforcolumnconfig", methods={"GET"})
-     */
+    #[Route("/get-metadata-for-column-config", name: "opendxp_admin_asset_assethelper_getmetadataforcolumnconfig", methods: ["GET"])]
     public function getMetadataForColumnConfigAction(Request $request): JsonResponse
     {
         $result = [];
@@ -932,9 +914,7 @@ class AssetHelperController extends AdminAbstractController
         return $this->adminJson($result);
     }
 
-    /**
-     * @Route("/get-batch-jobs", name="opendxp_admin_asset_assethelper_getbatchjobs", methods={"POST"})
-     */
+    #[Route("/get-batch-jobs", name: "opendxp_admin_asset_assethelper_getbatchjobs", methods: ["POST"])]
     public function getBatchJobsAction(Request $request, GridHelperService $gridHelperService): JsonResponse
     {
         if ($request->get('language')) {
@@ -949,9 +929,7 @@ class AssetHelperController extends AdminAbstractController
         return $this->adminJson(['success' => true, 'jobs' => $jobs]);
     }
 
-    /**
-     * @Route("/batch", name="opendxp_admin_asset_assethelper_batch", methods={"PUT"})
-     */
+    #[Route("/batch", name: "opendxp_admin_asset_assethelper_batch", methods: ["PUT"])]
     public function batchAction(Request $request, EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         try {
@@ -978,7 +956,7 @@ class AssetHelperController extends AdminAbstractController
                     $language = $data['language'] != 'default' ? $data['language'] : null;
                 }
 
-                $asset = Asset::getById($data['job']);
+                $asset = Asset::getById((int) $data['job']);
 
                 if ($asset) {
                     if (!$asset->isAllowed('publish')) {
