@@ -35,20 +35,17 @@ use Symfony\Component\HttpFoundation\Session\Attribute\AttributeBagInterface;
 use Symfony\Component\HttpKernel\Event\ControllerEvent;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
- * @Route("/class", name="opendxp_admin_dataobject_class_")
- *
  * @internal
  */
+#[Route("/class", name: "opendxp_admin_dataobject_class_")]
 class ClassController extends AdminAbstractController implements KernelControllerEventInterface
 {
-    /**
-     * @Route("/get-document-types", name="getdocumenttypes", methods={"GET"})
-     */
+    #[Route("/get-document-types", name: "getdocumenttypes", methods: ["GET"])]
     public function getDocumentTypesAction(Request $request): JsonResponse
     {
         $documentTypes = Document::getTypes();
@@ -62,9 +59,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson($typeItems);
     }
 
-    /**
-     * @Route("/get-asset-types", name="getassettypes", methods={"GET"})
-     */
+    #[Route("/get-asset-types", name: "getassettypes", methods: ["GET"])]
     public function getAssetTypesAction(Request $request): JsonResponse
     {
         $assetTypes = Asset::getTypes();
@@ -78,9 +73,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson($typeItems);
     }
 
-    /**
-     * @Route("/get-tree", name="gettree", methods={"GET", "POST"})
-     */
+    #[Route("/get-tree", name: "gettree", methods: ["GET", "POST"])]
     public function getTreeAction(Request $request): JsonResponse
     {
         try {
@@ -221,9 +214,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson($treeNodes);
     }
 
-    /**
-     * @Route("/get", name="get", methods={"GET"})
-     */
+    #[Route("/get", name: "get", methods: ["GET"])]
     public function getAction(Request $request): JsonResponse
     {
         $class = DataObject\ClassDefinition::getById($request->get('id'));
@@ -238,9 +229,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson($class);
     }
 
-    /**
-     * @Route("/get-custom-layout", name="getcustomlayout", methods={"GET"})
-     */
+    #[Route("/get-custom-layout", name: "getcustomlayout", methods: ["GET"])]
     public function getCustomLayoutAction(Request $request): JsonResponse
     {
         $customLayout = DataObject\ClassDefinition\CustomLayout::getById($request->get('id'));
@@ -276,9 +265,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson(['success' => true, 'data' => $customLayout]);
     }
 
-    /**
-     * @Route("/add", name="add", methods={"POST"})
-     */
+    #[Route("/add", name: "add", methods: ["POST"])]
     public function addAction(Request $request): JsonResponse
     {
         $className = $request->get('className');
@@ -302,9 +289,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson(['success' => true, 'id' => $class->getId()]);
     }
 
-    /**
-     * @Route("/add-custom-layout", name="addcustomlayout", methods={"POST"})
-     */
+    #[Route("/add-custom-layout", name: "addcustomlayout", methods: ["POST"])]
     public function addCustomLayoutAction(Request $request): JsonResponse
     {
         $layoutId = $request->get('layoutIdentifier');
@@ -335,9 +320,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
                                  'data' => $data, ]);
     }
 
-    /**
-     * @Route("/delete", name="delete", methods={"DELETE"})
-     */
+    #[Route("/delete", name: "delete", methods: ["DELETE"])]
     public function deleteAction(Request $request): Response
     {
         $class = DataObject\ClassDefinition::getById($request->get('id'));
@@ -348,9 +331,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return new Response();
     }
 
-    /**
-     * @Route("/delete-custom-layout", name="deletecustomlayout", methods={"DELETE"})
-     */
+    #[Route("/delete-custom-layout", name: "deletecustomlayout", methods: ["DELETE"])]
     public function deleteCustomLayoutAction(Request $request): JsonResponse
     {
         $customLayouts = new DataObject\ClassDefinition\CustomLayout\Listing();
@@ -368,9 +349,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/save-custom-layout", name="savecustomlayout", methods={"PUT"})
-     */
+    #[Route("/save-custom-layout", name: "savecustomlayout", methods: ["PUT"])]
     public function saveCustomLayoutAction(Request $request): JsonResponse
     {
         $customLayout = DataObject\ClassDefinition\CustomLayout::getById($request->get('id'));
@@ -410,10 +389,9 @@ class ClassController extends AdminAbstractController implements KernelControlle
     }
 
     /**
-     * @Route("/save", name="save", methods={"PUT"})
-     *
      * @throws \Exception
      */
+    #[Route("/save", name: "save", methods: ["PUT"])]
     public function saveAction(Request $request): JsonResponse
     {
         $class = DataObject\ClassDefinition::getById($request->get('id'));
@@ -501,9 +479,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $name;
     }
 
-    /**
-     * @Route("/import-class", name="importclass", methods={"POST", "PUT"})
-     */
+    #[Route("/import-class", name: "importclass", methods: ["POST", "PUT"])]
     public function importClassAction(Request $request): Response
     {
         $class = DataObject\ClassDefinition::getById($request->get('id'));
@@ -524,9 +500,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $response;
     }
 
-    /**
-     * @Route("/import-custom-layout-definition", name="importcustomlayoutdefinition", methods={"POST", "PUT"})
-     */
+    #[Route("/import-custom-layout-definition", name: "importcustomlayoutdefinition", methods: ["POST", "PUT"])]
     public function importCustomLayoutDefinitionAction(Request $request): Response
     {
         $success = false;
@@ -576,9 +550,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $response;
     }
 
-    /**
-     * @Route("/get-custom-layout-definitions", name="getcustomlayoutdefinitions", methods={"GET"})
-     */
+    #[Route("/get-custom-layout-definitions", name: "getcustomlayoutdefinitions", methods: ["GET"])]
     public function getCustomLayoutDefinitionsAction(Request $request): JsonResponse
     {
         $classIds = explode(',', $request->get('classId'));
@@ -600,9 +572,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson(['success' => true, 'data' => $result]);
     }
 
-    /**
-     * @Route("/get-all-layouts", name="getalllayouts", methods={"GET"})
-     */
+    #[Route("/get-all-layouts", name: "getalllayouts", methods: ["GET"])]
     public function getAllLayoutsAction(Request $request): JsonResponse
     {
         // get all classes
@@ -649,9 +619,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson(['data' => $resultList]);
     }
 
-    /**
-     * @Route("/export-class", name="exportclass", methods={"GET"})
-     */
+    #[Route("/export-class", name: "exportclass", methods: ["GET"])]
     public function exportClassAction(Request $request): Response
     {
         $id = $request->get('id');
@@ -668,14 +636,12 @@ class ClassController extends AdminAbstractController implements KernelControlle
 
         $response = new Response($json);
         $response->headers->set('Content-type', 'application/json');
-        $response->headers->set('Content-Disposition', 'attachment; filename="class_' . $class->getName() . '_export.json"');
+        $response->headers->set('Content-Disposition', 'attachment; filename: "class_' . $class->getName() . '_export.json"');
 
         return $response;
     }
 
-    /**
-     * @Route("/export-custom-layout-definition", name="exportcustomlayoutdefinition", methods={"GET"})
-     */
+    #[Route("/export-custom-layout-definition", name: "exportcustomlayoutdefinition", methods: ["GET"])]
     public function exportCustomLayoutDefinitionAction(Request $request): Response
     {
         $id = $request->get('id');
@@ -688,7 +654,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
 
                 $response = new Response($json);
                 $response->headers->set('Content-type', 'application/json');
-                $response->headers->set('Content-Disposition', 'attachment; filename="custom_definition_' . $name . '_export.json"');
+                $response->headers->set('Content-Disposition', 'attachment; filename: "custom_definition_' . $name . '_export.json"');
 
                 return $response;
             }
@@ -704,9 +670,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
      * FIELDCOLLECTIONS
      */
 
-    /**
-     * @Route("/fieldcollection-get", name="fieldcollectionget", methods={"GET"})
-     */
+    #[Route("/fieldcollection-get", name: "fieldcollectionget", methods: ["GET"])]
     public function fieldcollectionGetAction(Request $request): JsonResponse
     {
         $fc = DataObject\Fieldcollection\Definition::getByKey($request->get('id'));
@@ -718,9 +682,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson($fc);
     }
 
-    /**
-     * @Route("/fieldcollection-update", name="fieldcollectionupdate", methods={"PUT", "POST"})
-     */
+    #[Route("/fieldcollection-update", name: "fieldcollectionupdate", methods: ["PUT", "POST"])]
     public function fieldcollectionUpdateAction(Request $request): JsonResponse
     {
         try {
@@ -771,9 +733,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         }
     }
 
-    /**
-     * @Route("/import-fieldcollection", name="importfieldcollection", methods={"POST"})
-     */
+    #[Route("/import-fieldcollection", name: "importfieldcollection", methods: ["POST"])]
     public function importFieldcollectionAction(Request $request): Response
     {
         $this->checkPermission('fieldcollections');
@@ -795,9 +755,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $response;
     }
 
-    /**
-     * @Route("/export-fieldcollection", name="exportfieldcollection", methods={"GET"})
-     */
+    #[Route("/export-fieldcollection", name: "exportfieldcollection", methods: ["GET"])]
     public function exportFieldcollectionAction(Request $request): Response
     {
         $this->checkPermission('fieldcollections');
@@ -819,9 +777,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $response;
     }
 
-    /**
-     * @Route("/fieldcollection-delete", name="fieldcollectiondelete", methods={"DELETE"})
-     */
+    #[Route("/fieldcollection-delete", name: "fieldcollectiondelete", methods: ["DELETE"])]
     public function fieldcollectionDeleteAction(Request $request): JsonResponse
     {
         $this->checkPermission('fieldcollections');
@@ -832,9 +788,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/fieldcollection-tree", name="fieldcollectiontree", methods={"GET", "POST"})
-     */
+    #[Route("/fieldcollection-tree", name: "fieldcollectiontree", methods: ["GET", "POST"])]
     public function fieldcollectionTreeAction(Request $request, EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         $list = new DataObject\Fieldcollection\Definition\Listing();
@@ -934,9 +888,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson($definitions);
     }
 
-    /**
-     * @Route("/fieldcollection-list", name="fieldcollectionlist", methods={"GET"})
-     */
+    #[Route("/fieldcollection-list", name: "fieldcollectionlist", methods: ["GET"])]
     public function fieldcollectionListAction(Request $request, EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         $user = \OpenDxp\Tool\Admin::getCurrentUser();
@@ -983,9 +935,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson(['fieldcollections' => $list]);
     }
 
-    /**
-     * @Route("/get-class-definition-for-column-config", name="getclassdefinitionforcolumnconfig", methods={"GET"})
-     */
+    #[Route("/get-class-definition-for-column-config", name: "getclassdefinitionforcolumnconfig", methods: ["GET"])]
     public function getClassDefinitionForColumnConfigAction(Request $request): JsonResponse
     {
         $class = DataObject\ClassDefinition::getById($request->get('id'));
@@ -1061,9 +1011,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
      * OBJECT BRICKS
      */
 
-    /**
-     * @Route("/objectbrick-get", name="objectbrickget", methods={"GET"})
-     */
+    #[Route("/objectbrick-get", name: "objectbrickget", methods: ["GET"])]
     public function objectbrickGetAction(Request $request): JsonResponse
     {
         $fc = DataObject\Objectbrick\Definition::getByKey($request->get('id'));
@@ -1075,9 +1023,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson($fc);
     }
 
-    /**
-     * @Route("/objectbrick-update", name="objectbrickupdate", methods={"PUT", "POST"})
-     */
+    #[Route("/objectbrick-update", name: "objectbrickupdate", methods: ["PUT", "POST"])]
     public function objectbrickUpdateAction(Request $request, EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         try {
@@ -1137,9 +1083,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         }
     }
 
-    /**
-     * @Route("/import-objectbrick", name="importobjectbrick", methods={"POST"})
-     */
+    #[Route("/import-objectbrick", name: "importobjectbrick", methods: ["POST"])]
     public function importObjectbrickAction(Request $request): JsonResponse
     {
         $this->checkPermission('objectbricks');
@@ -1160,9 +1104,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $response;
     }
 
-    /**
-     * @Route("/export-objectbrick", name="exportobjectbrick", methods={"GET"})
-     */
+    #[Route("/export-objectbrick", name: "exportobjectbrick", methods: ["GET"])]
     public function exportObjectbrickAction(Request $request): Response
     {
         $this->checkPermission('objectbricks');
@@ -1184,9 +1126,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $response;
     }
 
-    /**
-     * @Route("/objectbrick-delete", name="objectbrickdelete", methods={"DELETE"})
-     */
+    #[Route("/objectbrick-delete", name: "objectbrickdelete", methods: ["DELETE"])]
     public function objectbrickDeleteAction(Request $request): JsonResponse
     {
         $this->checkPermission('objectbricks');
@@ -1197,9 +1137,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/objectbrick-tree", name="objectbricktree", methods={"GET", "POST"})
-     */
+    #[Route("/objectbrick-tree", name: "objectbricktree", methods: ["GET", "POST"])]
     public function objectbrickTreeAction(Request $request, EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         $list = new DataObject\Objectbrick\Definition\Listing();
@@ -1342,9 +1280,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         }
     }
 
-    /**
-     * @Route("/objectbrick-list", name="objectbricklist", methods={"GET"})
-     */
+    #[Route("/objectbrick-list", name: "objectbricklist", methods: ["GET"])]
     public function objectbrickListAction(Request $request, EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         $list = new DataObject\Objectbrick\Definition\Listing();
@@ -1408,9 +1344,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
      * Add option to export/import all class definitions/brick definitions etc. at once
      */
 
-    /**
-     * @Route("/bulk-import", name="bulkimport", methods={"POST"})
-     */
+    #[Route("/bulk-import", name: "bulkimport", methods: ["POST"])]
     public function bulkImportAction(Request $request): JsonResponse
     {
         $result = [];
@@ -1469,10 +1403,9 @@ class ClassController extends AdminAbstractController implements KernelControlle
      */
 
     /**
-     * @Route("/bulk-commit", name="bulkcommit", methods={"POST"})
-     *
      * @throws \Exception
      */
+    #[Route("/bulk-commit", name: "bulkcommit", methods: ["POST"])]
     public function bulkCommitAction(Request $request): JsonResponse
     {
         $data = json_decode($request->get('data'), true);
@@ -1574,9 +1507,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
      * Add option to export/import all class definitions/brick definitions etc. at once
      */
 
-    /**
-     * @Route("/bulk-export-prepare", name="bulkexportprepare", methods={"POST"})
-     */
+    #[Route("/bulk-export-prepare", name: "bulkexportprepare", methods: ["POST"])]
     public function bulkExportPrepareAction(Request $request): Response
     {
         $data = $request->get('data');
@@ -1588,9 +1519,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/bulk-export", name="bulkexport", methods={"GET"})
-     */
+    #[Route("/bulk-export", name: "bulkexport", methods: ["GET"])]
     public function bulkExportAction(Request $request): JsonResponse
     {
         $result = [];
@@ -1662,9 +1591,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return new JsonResponse(['success' => true, 'data' => $result]);
     }
 
-    /**
-     * @Route("/do-bulk-export", name="dobulkexport", methods={"GET"})
-     */
+    #[Route("/do-bulk-export", name: "dobulkexport", methods: ["GET"])]
     public function doBulkExportAction(Request $request): Response
     {
         $session = Session::getSessionBag($request->getSession(), 'opendxp_objects');
@@ -1728,9 +1655,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         $this->checkActionPermission($event, 'classes', $unrestrictedActions);
     }
 
-    /**
-     * @Route("/get-fieldcollection-usages", name="getfieldcollectionusages", methods={"GET"})
-     */
+    #[Route("/get-fieldcollection-usages", name: "getfieldcollectionusages", methods: ["GET"])]
     public function getFieldcollectionUsagesAction(Request $request): Response
     {
         $key = $request->get('key');
@@ -1756,9 +1681,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson($result);
     }
 
-    /**
-     * @Route("/get-bricks-usages", name="getbrickusages", methods={"GET"})
-     */
+    #[Route("/get-bricks-usages", name: "getbrickusages", methods: ["GET"])]
     public function getBrickUsagesAction(Request $request): Response
     {
         $classId = $request->get('classId');
@@ -1801,9 +1724,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson($usages);
     }
 
-    /**
-     * @Route("/get-icons", name="geticons", methods={"GET"})
-     */
+    #[Route("/get-icons", name: "geticons", methods: ["GET"])]
     public function getIconsAction(Request $request, EventDispatcherInterface $eventDispatcher): Response
     {
         $classId = $request->query->get('classId');
@@ -1898,9 +1819,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson($result);
     }
 
-    /**
-     * @Route("/suggest-class-identifier", name="suggestclassidentifier")
-     */
+    #[Route("/suggest-class-identifier", name: "suggestclassidentifier")]
     public function suggestClassIdentifierAction(): Response
     {
         $db = Db::get();
@@ -1916,9 +1835,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson($result);
     }
 
-    /**
-     * @Route("/suggest-custom-layout-identifier", name="suggestcustomlayoutidentifier")
-     */
+    #[Route("/suggest-custom-layout-identifier", name: "suggestcustomlayoutidentifier")]
     public function suggestCustomLayoutIdentifierAction(Request $request): Response
     {
         $classId = $request->get('classId');
@@ -1947,9 +1864,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $this->adminJson($result);
     }
 
-    /**
-     * @Route("/text-layout-preview", name="textlayoutpreview")
-     */
+    #[Route("/text-layout-preview", name: "textlayoutpreview")]
     public function textLayoutPreviewAction(Request $request): Response
     {
         $objPath = $request->get('previewObject', '');
@@ -1992,9 +1907,7 @@ class ClassController extends AdminAbstractController implements KernelControlle
         return $response;
     }
 
-    /**
-     * @Route("/video-supported-types", name="videosupportedTypestypes")
-     */
+    #[Route("/video-supported-types", name: "videosupportedTypestypes")]
     public function videoAllowedTypesAction(Request $request, TranslatorInterface $translator): Response
     {
         $videoDef = new DataObject\ClassDefinition\Data\Video();
