@@ -48,16 +48,15 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Session\Attribute\AttributeBagInterface;
 use Symfony\Component\HttpKernel\Event\ControllerEvent;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 use Twig\Environment;
 use Twig\Extension\CoreExtension;
 
 /**
- * @Route("/object", name="opendxp_admin_dataobject_dataobject_")
- *
  * @internal
  */
+#[Route("/object", name: "opendxp_admin_dataobject_dataobject_")]
 class DataObjectController extends ElementControllerBase implements KernelControllerEventInterface
 {
     use AdminStyleTrait;
@@ -83,9 +82,7 @@ class DataObjectController extends ElementControllerBase implements KernelContro
 
     private array $classFieldDefinitions = [];
 
-    /**
-     * @Route("/tree-get-children-by-id", name="treegetchildrenbyid", methods={"GET"})
-     */
+    #[Route("/tree-get-children-by-id", name: "treegetchildrenbyid", methods: ["GET"])]
     public function treeGetChildrenByIdAction(Request $request, EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         $allParams = array_merge($request->request->all(), $request->query->all());
@@ -240,9 +237,7 @@ class DataObjectController extends ElementControllerBase implements KernelContro
         return $this->elementService->getElementTreeNodeConfig($element);
     }
 
-    /**
-     * @Route("/get-id-path-paging-info", name="getidpathpaginginfo", methods={"GET"})
-     */
+    #[Route("/get-id-path-paging-info", name: "getidpathpaginginfo", methods: ["GET"])]
     public function getIdPathPagingInfoAction(Request $request): JsonResponse
     {
         $path = $request->get('path');
@@ -286,10 +281,9 @@ class DataObjectController extends ElementControllerBase implements KernelContro
     }
 
     /**
-     * @Route("/get", name="get", methods={"GET"})
-     *
      * @throws \Exception
      */
+    #[Route("/get", name: "get", methods: ["GET"])]
     public function getAction(Request $request, EventDispatcherInterface $eventDispatcher, PreviewGeneratorInterface $defaultPreviewGenerator): JsonResponse
     {
         $objectId = $request->query->getInt('id');
@@ -548,10 +542,9 @@ class DataObjectController extends ElementControllerBase implements KernelContro
     }
 
     /**
-     * @Route("/get-select-options", name="getSelectOptions", methods={"POST"})
-     *
      * @throws \Exception
      */
+    #[Route("/get-select-options", name: "getSelectOptions", methods: ["POST"])]
     public function getSelectOptions(Request $request): JsonResponse
     {
         $objectId = $request->request->getInt('objectId');
@@ -751,9 +744,7 @@ class DataObjectController extends ElementControllerBase implements KernelContro
         }
     }
 
-    /**
-     * @Route("/get-folder", name="getfolder", methods={"GET"})
-     */
+    #[Route("/get-folder", name: "getfolder", methods: ["GET"])]
     public function getFolderAction(Request $request, EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         $objectId = (int)$request->get('id');
@@ -839,9 +830,7 @@ class DataObjectController extends ElementControllerBase implements KernelContro
         return $reduced;
     }
 
-    /**
-     * @Route("/add", name="add", methods={"POST"})
-     */
+    #[Route("/add", name: "add", methods: ["POST"])]
     public function addAction(Request $request, Model\FactoryInterface $modelFactory): JsonResponse
     {
         $message = '';
@@ -909,9 +898,7 @@ class DataObjectController extends ElementControllerBase implements KernelContro
         return $this->adminJson($return);
     }
 
-    /**
-     * @Route("/add-folder", name="addfolder", methods={"POST"})
-     */
+    #[Route("/add-folder", name: "addfolder", methods: ["POST"])]
     public function addFolderAction(Request $request): JsonResponse
     {
         $success = false;
@@ -943,10 +930,9 @@ class DataObjectController extends ElementControllerBase implements KernelContro
     }
 
     /**
-     * @Route("/delete", name="delete", methods={"DELETE"})
-     *
      * @throws \Exception
      */
+    #[Route("/delete", name: "delete", methods: ["DELETE"])]
     public function deleteAction(Request $request): JsonResponse
     {
         $type = $request->get('type');
@@ -990,10 +976,9 @@ class DataObjectController extends ElementControllerBase implements KernelContro
     }
 
     /**
-     * @Route("/change-children-sort-by", name="changechildrensortby", methods={"PUT"})
-     *
      * @throws \Exception
      */
+    #[Route("/change-children-sort-by", name: "changechildrensortby", methods: ["PUT"])]
     public function changeChildrenSortByAction(Request $request): JsonResponse
     {
         $object = DataObject::getById((int) $request->get('id'));
@@ -1030,10 +1015,9 @@ class DataObjectController extends ElementControllerBase implements KernelContro
     }
 
     /**
-     * @Route("/update", name="update", methods={"PUT"})
-     *
      * @throws \Exception
      */
+    #[Route("/update", name: "update", methods: ["PUT"])]
     public function updateAction(Request $request): JsonResponse
     {
         $values = $this->decodeJson($request->get('values'));
@@ -1094,7 +1078,7 @@ class DataObjectController extends ElementControllerBase implements KernelContro
             }
 
             if (!empty($values['parentId'])) {
-                $parent = DataObject::getById($values['parentId']);
+                $parent = DataObject::getById((int) $values['parentId']);
 
                 //check if parent is changed
                 if ($object->getParentId() != $parent->getId()) {
@@ -1310,10 +1294,9 @@ class DataObjectController extends ElementControllerBase implements KernelContro
     }
 
     /**
-     * @Route("/save", name="save", methods={"POST", "PUT"})
-     *
      * @throws \Exception
      */
+    #[Route("/save", name: "save", methods: ["POST", "PUT"])]
     public function saveAction(Request $request): JsonResponse
     {
         $objectFromDatabase = DataObject\Concrete::getById((int) $request->get('id'));
@@ -1482,9 +1465,7 @@ class DataObjectController extends ElementControllerBase implements KernelContro
         return true;
     }
 
-    /**
-     * @Route("/save-folder", name="savefolder", methods={"PUT"})
-     */
+    #[Route("/save-folder", name: "savefolder", methods: ["PUT"])]
     public function saveFolderAction(Request $request): JsonResponse
     {
         $object = DataObject::getById((int) $request->get('id'));
@@ -1548,9 +1529,7 @@ class DataObjectController extends ElementControllerBase implements KernelContro
         }
     }
 
-    /**
-     * @Route("/publish-version", name="publishversion", methods={"POST"})
-     */
+    #[Route("/publish-version", name: "publishversion", methods: ["POST"])]
     public function publishVersionAction(Request $request): JsonResponse
     {
         $id = (int)$request->get('id');
@@ -1587,10 +1566,9 @@ class DataObjectController extends ElementControllerBase implements KernelContro
     }
 
     /**
-     * @Route("/preview-version", name="previewversion", methods={"GET"})
-     *
      * @throws \Exception
      */
+    #[Route("/preview-version", name: "previewversion", methods: ["GET"])]
     public function previewVersionAction(Request $request, Environment $twig): Response
     {
         DataObject::setDoNotRestoreKeyAndPath(true);
@@ -1631,10 +1609,9 @@ class DataObjectController extends ElementControllerBase implements KernelContro
     }
 
     /**
-     * @Route("/diff-versions/from/{from}/to/{to}", name="diffversions", methods={"GET"})
-     *
      * @throws \Exception
      */
+    #[Route("/diff-versions/from/{from}/to/{to}", name: "diffversions", methods: ["GET"])]
     public function diffVersionsAction(Request $request, Environment $twig, int $from, int $to): Response
     {
         DataObject::setDoNotRestoreKeyAndPath(true);
@@ -1690,9 +1667,7 @@ class DataObjectController extends ElementControllerBase implements KernelContro
         throw $this->createAccessDeniedException('Permission denied, version ids [' . $id1 . ', ' . $id2 . ']');
     }
 
-    /**
-     * @Route("/grid-proxy", name="gridproxy", methods={"GET", "POST", "PUT"})
-     */
+    #[Route("/grid-proxy", name: "gridproxy", methods: ["GET", "POST", "PUT"])]
     public function gridProxyAction(
         Request $request,
         EventDispatcherInterface $eventDispatcher,
@@ -1728,9 +1703,7 @@ class DataObjectController extends ElementControllerBase implements KernelContro
         return $this->adminJson($result);
     }
 
-    /**
-     * @Route("/copy-info", name="copyinfo", methods={"GET"})
-     */
+    #[Route("/copy-info", name: "copyinfo", methods: ["GET"])]
     public function copyInfoAction(Request $request): JsonResponse
     {
         $transactionId = time();
@@ -1815,10 +1788,9 @@ class DataObjectController extends ElementControllerBase implements KernelContro
     }
 
     /**
-     * @Route("/copy-rewrite-ids", name="copyrewriteids", methods={"PUT"})
-     *
      * @throws \Exception
      */
+    #[Route("/copy-rewrite-ids", name: "copyrewriteids", methods: ["PUT"])]
     public function copyRewriteIdsAction(Request $request): JsonResponse
     {
         $transactionId = $request->get('transactionId');
@@ -1853,9 +1825,7 @@ class DataObjectController extends ElementControllerBase implements KernelContro
         ]);
     }
 
-    /**
-     * @Route("/copy", name="copy", methods={"POST"})
-     */
+    #[Route("/copy", name: "copy", methods: ["POST"])]
     public function copyAction(Request $request): JsonResponse
     {
         $message = '';
@@ -1871,7 +1841,7 @@ class DataObjectController extends ElementControllerBase implements KernelContro
 
             // this is because the key can get the prefix "_copy" if the target does already exists
             if ($sessionBag['parentId']) {
-                $targetParent = DataObject::getById($sessionBag['parentId']);
+                $targetParent = DataObject::getById((int) $sessionBag['parentId']);
             } else {
                 $targetParent = DataObject::getById((int) $request->get('targetParentId'));
             }
@@ -1919,9 +1889,7 @@ class DataObjectController extends ElementControllerBase implements KernelContro
         }
     }
 
-    /**
-     * @Route("/preview", name="preview", methods={"GET"})
-     */
+    #[Route("/preview", name: "preview", methods: ["GET"])]
     public function previewAction(Request $request, PreviewGeneratorInterface $defaultPreviewGenerator): RedirectResponse|Response
     {
         $id = $request->query->getInt('id');
