@@ -38,23 +38,21 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Messenger\MessageBusInterface;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 use Twig\Environment;
 
 /**
- * @Route("/page", name="opendxp_admin_document_page_")
- *
  * @internal
  */
+#[Route("/page", name: "opendxp_admin_document_page_")]
 class PageController extends DocumentControllerBase
 {
     use RecursionBlockingEventDispatchHelperTrait;
 
     /**
-     * @Route("/get-data-by-id", name="getdatabyid", methods={"GET"})
-     *
      * @throws \Exception
      */
+    #[Route("/get-data-by-id", name: "getdatabyid", methods: ["GET"])]
     public function getDataByIdAction(Request $request, StaticPageGenerator $staticPageGenerator): JsonResponse
     {
         $page = Document\Page::getById((int)$request->get('id'));
@@ -106,10 +104,9 @@ class PageController extends DocumentControllerBase
     }
 
     /**
-     * @Route("/save", name="save", methods={"PUT", "POST"})
-     *
      * @throws \Exception
      */
+    #[Route("/save", name: "save", methods: ["PUT", "POST"])]
     public function saveAction(Request $request, StaticPageGenerator $staticPageGenerator): JsonResponse
     {
         $oldPage = Document\Page::getById((int) $request->get('id'));
@@ -181,9 +178,7 @@ class PageController extends DocumentControllerBase
         }
     }
 
-    /**
-     * @Route("/generate-previews", name="generatepreviews", methods={"GET"})
-     */
+    #[Route("/generate-previews", name: "generatepreviews", methods: ["GET"])]
     public function generatePreviewsAction(Request $request, MessageBusInterface $messengerBusOpendxpCore): JsonResponse
     {
         $list = new Document\Listing();
@@ -200,9 +195,7 @@ class PageController extends DocumentControllerBase
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/display-preview-image", name="display_preview_image", methods={"GET"})
-     */
+    #[Route("/display-preview-image", name: "display_preview_image", methods: ["GET"])]
     public function displayPreviewImageAction(Request $request): BinaryFileResponse
     {
         $document = Document\Page::getById((int) $request->get('id'));
@@ -215,9 +208,7 @@ class PageController extends DocumentControllerBase
         throw $this->createNotFoundException('Page not found');
     }
 
-    /**
-     * @Route("/check-pretty-url", name="checkprettyurl", methods={"POST"})
-     */
+    #[Route("/check-pretty-url", name: "checkprettyurl", methods: ["POST"])]
     public function checkPrettyUrlAction(Request $request): JsonResponse
     {
         $docId = $request->request->getInt('id');
@@ -284,9 +275,7 @@ class PageController extends DocumentControllerBase
         ]);
     }
 
-    /**
-     * @Route("/clear-editable-data", name="cleareditabledata", methods={"PUT"})
-     */
+    #[Route("/clear-editable-data", name: "cleareditabledata", methods: ["PUT"])]
     public function clearEditableDataAction(Request $request): JsonResponse
     {
         $docId = $request->request->getInt('id');
@@ -312,10 +301,9 @@ class PageController extends DocumentControllerBase
     }
 
     /**
-     * @Route("/qr-code", name="qrcode", methods={"GET"})
-     *
      * @throws \Exception
      */
+    #[Route("/qr-code", name: "qrcode", methods: ["GET"])]
     public function qrCodeAction(Request $request): BinaryFileResponse
     {
         $page = Document\Page::getById((int) $request->query->get('id'));
@@ -348,10 +336,9 @@ class PageController extends DocumentControllerBase
     }
 
     /**
-     * @Route("/areabrick-render-index-editmode", name="areabrick-render-index-editmode", methods={"POST"})
-     *
      * @throws NotFoundHttpException|\Exception
      */
+    #[Route("/areabrick-render-index-editmode", name: "areabrick-render-index-editmode", methods: ["POST"])]
     public function areabrickRenderIndexEditmode(
         Request $request,
         BlockStateStack $blockStateStack,
