@@ -31,7 +31,7 @@ use OpenDxp\Model\Version;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
@@ -40,9 +40,7 @@ use Symfony\Contracts\Translation\TranslatorInterface;
  */
 class ElementController extends AdminAbstractController
 {
-    /**
-     * @Route("/element/lock-element", name="opendxp_admin_element_lockelement", methods={"PUT"})
-     */
+    #[Route("/element/lock-element", name: "opendxp_admin_element_lockelement", methods: ["PUT"])]
     public function lockElementAction(Request $request): Response
     {
         Element\Editlock::lock($request->request->getInt('id'), $request->request->get('type'), $request->getSession()->getId());
@@ -50,9 +48,7 @@ class ElementController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/element/unlock-element", name="opendxp_admin_element_unlockelement", methods={"PUT"})
-     */
+    #[Route("/element/unlock-element", name: "opendxp_admin_element_unlockelement", methods: ["PUT"])]
     public function unlockElementAction(Request $request): Response
     {
         Element\Editlock::unlock((int)$request->get('id'), $request->get('type'));
@@ -60,9 +56,7 @@ class ElementController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/element/unlock-elements", name="opendxp_admin_element_unlockelements", methods={"POST"})
-     */
+    #[Route("/element/unlock-elements", name: "opendxp_admin_element_unlockelements", methods: ["POST"])]
     public function unlockElementsAction(Request $request): Response
     {
         $request = json_decode($request->getContent(), true) ?? [];
@@ -75,9 +69,8 @@ class ElementController extends AdminAbstractController
 
     /**
      * Returns the element data denoted by the given type and ID or path.
-     *
-     * @Route("/element/get-subtype", name="opendxp_admin_element_getsubtype", methods={"GET"})
      */
+    #[Route("/element/get-subtype", name: "opendxp_admin_element_getsubtype", methods: ["GET"])]
     public function getSubtypeAction(Request $request): JsonResponse
     {
         $idOrPath = trim($request->query->get('id', ''));
@@ -134,9 +127,7 @@ class ElementController extends AdminAbstractController
         return $this->adminJson(['noteTypes' => $result]);
     }
 
-    /**
-     * @Route("/element/note-types", name="opendxp_admin_element_notetypes", methods={"GET"})
-     */
+    #[Route("/element/note-types", name: "opendxp_admin_element_notetypes", methods: ["GET"])]
     public function noteTypes(Request $request): JsonResponse
     {
         switch ($request->get('ctype')) {
@@ -151,9 +142,7 @@ class ElementController extends AdminAbstractController
         }
     }
 
-    /**
-     * @Route("/element/note-list", name="opendxp_admin_element_notelist", methods={"POST"})
-     */
+    #[Route("/element/note-list", name: "opendxp_admin_element_notelist", methods: ["POST"])]
     public function noteListAction(Request $request): JsonResponse
     {
         $this->checkPermission('notes_events');
@@ -280,9 +269,7 @@ class ElementController extends AdminAbstractController
         ]);
     }
 
-    /**
-     * @Route("/element/note-add", name="opendxp_admin_element_noteadd", methods={"POST"})
-     */
+    #[Route("/element/note-add", name: "opendxp_admin_element_noteadd", methods: ["POST"])]
     public function noteAddAction(Request $request): JsonResponse
     {
         $this->checkPermission('notes_events');
@@ -302,9 +289,7 @@ class ElementController extends AdminAbstractController
         ]);
     }
 
-    /**
-     * @Route("/element/find-usages", name="opendxp_admin_element_findusages", methods={"GET"})
-     */
+    #[Route("/element/find-usages", name: "opendxp_admin_element_findusages", methods: ["GET"])]
     public function findUsagesAction(Request $request): JsonResponse
     {
         $element = null;
@@ -341,7 +326,7 @@ class ElementController extends AdminAbstractController
                     ->getRequiredByWithPath($queryOffset, $queryLimit, $orderBy, $orderDirection);
 
                 foreach ($elements as $el) {
-                    $item = Element\Service::getElementById($el['type'], $el['id']);
+                    $item = Element\Service::getElementById($el['type'], (int) $el['id']);
 
                     if ($item instanceof Element\ElementInterface) {
                         if ($item->isAllowed('list')) {
@@ -367,9 +352,7 @@ class ElementController extends AdminAbstractController
         ]);
     }
 
-    /**
-     * @Route("/element/get-replace-assignments-batch-jobs", name="opendxp_admin_element_getreplaceassignmentsbatchjobs", methods={"GET"})
-     */
+    #[Route("/element/get-replace-assignments-batch-jobs", name: "opendxp_admin_element_getreplaceassignmentsbatchjobs", methods: ["GET"])]
     public function getReplaceAssignmentsBatchJobsAction(Request $request): JsonResponse
     {
         $element = null;
@@ -390,9 +373,7 @@ class ElementController extends AdminAbstractController
         }
     }
 
-    /**
-     * @Route("/element/replace-assignments", name="opendxp_admin_element_replaceassignments", methods={"POST"})
-     */
+    #[Route("/element/replace-assignments", name: "opendxp_admin_element_replaceassignments", methods: ["POST"])]
     public function replaceAssignmentsAction(Request $request): JsonResponse
     {
         $success = false;
@@ -434,9 +415,7 @@ class ElementController extends AdminAbstractController
         ]);
     }
 
-    /**
-     * @Route("/element/unlock-propagate", name="opendxp_admin_element_unlockpropagate", methods={"PUT"})
-     */
+    #[Route("/element/unlock-propagate", name: "opendxp_admin_element_unlockpropagate", methods: ["PUT"])]
     public function unlockPropagateAction(Request $request): JsonResponse
     {
         $success = false;
@@ -452,9 +431,7 @@ class ElementController extends AdminAbstractController
         ]);
     }
 
-    /**
-     * @Route("/element/type-path", name="opendxp_admin_element_typepath", methods={"GET"})
-     */
+    #[Route("/element/type-path", name: "opendxp_admin_element_typepath", methods: ["GET"])]
     public function typePathAction(Request $request): JsonResponse
     {
         $id = $request->query->getInt('id');
@@ -491,9 +468,7 @@ class ElementController extends AdminAbstractController
         return $this->adminJson($data);
     }
 
-    /**
-     * @Route("/element/version-update", name="opendxp_admin_element_versionupdate", methods={"PUT"})
-     */
+    #[Route("/element/version-update", name: "opendxp_admin_element_versionupdate", methods: ["PUT"])]
     public function versionUpdateAction(Request $request): JsonResponse
     {
         $data = $this->decodeJson($request->get('data'));
@@ -510,10 +485,9 @@ class ElementController extends AdminAbstractController
     }
 
     /**
-     * @Route("/element/get-nice-path", name="opendxp_admin_element_getnicepath", methods={"POST"})
-     *
      * @throws \Exception
      */
+    #[Route("/element/get-nice-path", name: "opendxp_admin_element_getnicepath", methods: ["POST"])]
     public function getNicePathAction(Request $request): JsonResponse
     {
         $source = $this->decodeJson($request->get('source'));
@@ -565,10 +539,9 @@ class ElementController extends AdminAbstractController
     }
 
     /**
-     * @Route("/element/get-versions", name="opendxp_admin_element_getversions", methods={"GET"})
-     *
      * @throws \Exception
      */
+    #[Route("/element/get-versions", name: "opendxp_admin_element_getversions", methods: ["GET"])]
     public function getVersionsAction(Request $request): JsonResponse
     {
         $id = (int)$request->get('id');
@@ -621,9 +594,7 @@ class ElementController extends AdminAbstractController
         throw $this->createNotFoundException('Element type not found');
     }
 
-    /**
-     * @Route("/element/delete-draft", name="opendxp_admin_element_deletedraft", methods={"DELETE"})
-     */
+    #[Route("/element/delete-draft", name: "opendxp_admin_element_deletedraft", methods: ["DELETE"])]
     public function deleteDraftAction(Request $request): JsonResponse
     {
         $version = Version::getById((int) $request->get('id'));
@@ -634,9 +605,7 @@ class ElementController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/element/delete-version", name="opendxp_admin_element_deleteversion", methods={"DELETE"})
-     */
+    #[Route("/element/delete-version", name: "opendxp_admin_element_deleteversion", methods: ["DELETE"])]
     public function deleteVersionAction(Request $request): JsonResponse
     {
         $version = Model\Version::getById((int) $request->get('id'));
@@ -645,9 +614,7 @@ class ElementController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/element/delete-all-versions", name="opendxp_admin_element_deleteallversion", methods={"DELETE"})
-     */
+    #[Route("/element/delete-all-versions", name: "opendxp_admin_element_deleteallversion", methods: ["DELETE"])]
     public function deleteAllVersionAction(Request $request): JsonResponse
     {
         $elementId = $request->request->getInt('id');
@@ -666,9 +633,7 @@ class ElementController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/element/get-requires-dependencies", name="opendxp_admin_element_getrequiresdependencies", methods={"GET"})
-     */
+    #[Route("/element/get-requires-dependencies", name: "opendxp_admin_element_getrequiresdependencies", methods: ["GET"])]
     public function getRequiresDependenciesAction(Request $request): JsonResponse
     {
         $id = $request->query->getInt('id');
@@ -728,9 +693,7 @@ class ElementController extends AdminAbstractController
         return $this->adminJson(false);
     }
 
-    /**
-     * @Route("/element/get-required-by-dependencies", name="opendxp_admin_element_getrequiredbydependencies", methods={"GET"})
-     */
+    #[Route("/element/get-required-by-dependencies", name: "opendxp_admin_element_getrequiredbydependencies", methods: ["GET"])]
     public function getRequiredByDependenciesAction(Request $request): JsonResponse
     {
         $id = $request->query->getInt('id');
@@ -790,9 +753,7 @@ class ElementController extends AdminAbstractController
         return $this->adminJson(false);
     }
 
-    /**
-     * @Route("/element/get-predefined-properties", name="opendxp_admin_element_getpredefinedproperties", methods={"GET"})
-     */
+    #[Route("/element/get-predefined-properties", name: "opendxp_admin_element_getpredefinedproperties", methods: ["GET"])]
     public function getPredefinedPropertiesAction(Request $request, TranslatorInterface $translator): JsonResponse
     {
         $properties = [];
@@ -821,9 +782,7 @@ class ElementController extends AdminAbstractController
         return $this->adminJson(['properties' => $properties]);
     }
 
-    /**
-     * @Route("/element/analyze-permissions", name="opendxp_admin_element_analyzepermissions", methods={"POST"})
-     */
+    #[Route("/element/analyze-permissions", name: "opendxp_admin_element_analyzepermissions", methods: ["POST"])]
     public function analyzePermissionsAction(Request $request): Response
     {
         $userId = $request->request->getInt('userId');
