@@ -22,18 +22,15 @@ use OpenDxp\Tool\Requirements;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Profiler\Profiler;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 
 /**
- * @Route("/install")
- *
  * @internal
  */
+#[Route("/install")]
 class InstallController extends AdminAbstractController
 {
-    /**
-     * @Route("/check", name="opendxp_admin_install_check", methods={"GET", "POST"})
-     */
+    #[Route("/check", name: "opendxp_admin_install_check", methods: ["GET", "POST"])]
     public function checkAction(Request $request, Connection $db, ?Profiler $profiler): Response
     {
         if ($profiler) {
