@@ -25,19 +25,18 @@ use OpenDxp\Model\Notification\Service\UserService;
 use OpenDxp\Model\User;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
- * @Route("/notification")
  *
  * @internal
  */
+
+#[Route("/notification")]
 class NotificationController extends AdminAbstractController
 {
-    /**
-     * @Route("/recipients", name="opendxp_admin_notification_recipients", methods={"GET"})
-     */
+    #[Route("/recipients", name: "opendxp_admin_notification_recipients", methods: ["GET"])]
     public function recipientsAction(UserService $service, TranslatorInterface $translator): JsonResponse
     {
         $this->checkPermission('notifications_send');
@@ -57,9 +56,7 @@ class NotificationController extends AdminAbstractController
         return $this->adminJson($data);
     }
 
-    /**
-     * @Route("/send", name="opendxp_admin_notification_send", methods={"POST"})
-     */
+    #[Route("/send", name: "opendxp_admin_notification_send", methods: ["POST"])]
     public function sendAction(Request $request, NotificationService $service): JsonResponse
     {
         $this->checkPermission('notifications_send');
@@ -85,9 +82,7 @@ class NotificationController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/find", name="opendxp_admin_notification_find", methods={"GET"})
-     */
+    #[Route("/find", name: "opendxp_admin_notification_find", methods: ["GET"])]
     public function findAction(Request $request, NotificationService $service): JsonResponse
     {
         $this->checkPermission('notifications');
@@ -112,9 +107,7 @@ class NotificationController extends AdminAbstractController
         ]);
     }
 
-    /**
-     * @Route("/find-all", name="opendxp_admin_notification_findall", methods={"POST"})
-     */
+    #[Route("/find-all", name: "opendxp_admin_notification_findall", methods: ["POST"])]
     public function findAllAction(Request $request, NotificationService $service): JsonResponse
     {
         $this->checkPermission('notifications');
@@ -146,9 +139,7 @@ class NotificationController extends AdminAbstractController
         ]);
     }
 
-    /**
-     * @Route("/find-last-unread", name="opendxp_admin_notification_findlastunread", methods={"GET"})
-     */
+    #[Route("/find-last-unread", name: "opendxp_admin_notification_findlastunread", methods: ["GET"])]
     public function findLastUnreadAction(Request $request, NotificationService $service): JsonResponse
     {
         $this->checkPermission('notifications');
@@ -172,9 +163,7 @@ class NotificationController extends AdminAbstractController
         ]);
     }
 
-    /**
-     * @Route("/mark-as-read", name="opendxp_admin_notification_markasread", methods={"PUT"})
-     */
+    #[Route("/mark-as-read", name: "opendxp_admin_notification_markasread", methods: ["PUT"])]
     public function markAsReadAction(Request $request, NotificationService $service): JsonResponse
     {
         $this->checkPermission('notifications');
@@ -185,9 +174,7 @@ class NotificationController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/delete", name="opendxp_admin_notification_delete", methods={"DELETE"})
-     */
+    #[Route("/delete", name: "opendxp_admin_notification_delete", methods: ["DELETE"])]
     public function deleteAction(Request $request, NotificationService $service): JsonResponse
     {
         $this->checkPermission('notifications');
@@ -198,9 +185,7 @@ class NotificationController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/delete-all", name="opendxp_admin_notification_deleteall", methods={"DELETE"})
-     */
+    #[Route("/delete-all", name: "opendxp_admin_notification_deleteall", methods: ["DELETE"])]
     public function deleteAllAction(NotificationService $service): JsonResponse
     {
         $this->checkPermission('notifications');
