@@ -47,14 +47,13 @@ use Symfony\Component\HttpKernel\Event\TerminateEvent;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\HttpKernel\KernelEvents;
 use Symfony\Component\HttpKernel\KernelInterface;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
- * @Route("/settings")
- *
  * @internal
  */
+#[Route("/settings")]
 class SettingsController extends AdminAbstractController
 {
     use StopMessengerWorkersTrait;
@@ -65,9 +64,7 @@ class SettingsController extends AdminAbstractController
     {
     }
 
-    /**
-     * @Route("/display-custom-logo", name="opendxp_settings_display_custom_logo", methods={"GET"})
-     */
+    #[Route("/display-custom-logo", name: "opendxp_settings_display_custom_logo", methods: ["GET"])]
     public function displayCustomLogoAction(Request $request): StreamedResponse
     {
         $mime = 'image/svg+xml';
@@ -98,10 +95,9 @@ class SettingsController extends AdminAbstractController
     }
 
     /**
-     * @Route("/upload-custom-logo", name="opendxp_admin_settings_uploadcustomlogo", methods={"POST"})
-     *
      * @throws \Exception
      */
+    #[Route("/upload-custom-logo", name: "opendxp_admin_settings_uploadcustomlogo", methods: ["POST"])]
     public function uploadCustomLogoAction(Request $request): JsonResponse
     {
         $logoFile = $request->files->get('Filedata');
@@ -124,9 +120,7 @@ class SettingsController extends AdminAbstractController
         return $response;
     }
 
-    /**
-     * @Route("/delete-custom-logo", name="opendxp_admin_settings_deletecustomlogo", methods={"DELETE"})
-     */
+    #[Route("/delete-custom-logo", name: "opendxp_admin_settings_deletecustomlogo", methods: ["DELETE"])]
     public function deleteCustomLogoAction(Request $request): JsonResponse
     {
         if (Tool\Storage::get('admin')->fileExists(self::CUSTOM_LOGO_PATH)) {
@@ -138,9 +132,8 @@ class SettingsController extends AdminAbstractController
 
     /**
      * Used by the predefined metadata grid
-     *
-     * @Route("/predefined-metadata", name="opendxp_admin_settings_metadata", methods={"POST"})
      */
+    #[Route("/predefined-metadata", name: "opendxp_admin_settings_metadata", methods: ["POST"])]
     public function metadataAction(Request $request): JsonResponse
     {
         $this->checkPermission('asset_metadata');
@@ -233,9 +226,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson(['success' => false]);
     }
 
-    /**
-     * @Route("/get-predefined-metadata", name="opendxp_admin_settings_getpredefinedmetadata", methods={"GET"})
-     */
+    #[Route("/get-predefined-metadata", name: "opendxp_admin_settings_getpredefinedmetadata", methods: ["GET"])]
     public function getPredefinedMetadataAction(Request $request): JsonResponse
     {
         $type = $request->get('type');
@@ -256,9 +247,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson(['data' => $result, 'success' => true]);
     }
 
-    /**
-     * @Route("/properties", name="opendxp_admin_settings_properties", methods={"POST"})
-     */
+    #[Route("/properties", name: "opendxp_admin_settings_properties", methods: ["POST"])]
     public function propertiesAction(Request $request): JsonResponse
     {
         if ($request->get('data')) {
@@ -346,9 +335,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson(['success' => false]);
     }
 
-    /**
-     * @Route("/get-admin-system", name="opendxp_appearance_admin_settings_get", methods={"GET"})
-     */
+    #[Route("/get-admin-system", name: "opendxp_appearance_admin_settings_get", methods: ["GET"])]
     public function getAppearanceSystemAction(AdminConfig $config): JsonResponse
     {
         $this->checkPermission('system_appearance_settings');
@@ -361,9 +348,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson($response);
     }
 
-    /**
-     * @Route("/get-system", name="opendxp_admin_settings_getsystem", methods={"GET"})
-     */
+    #[Route("/get-system", name: "opendxp_admin_settings_getsystem", methods: ["GET"])]
     public function getSystemAction(Request $request, SystemSettingsConfig $config): JsonResponse
     {
         $this->checkPermission('system_settings');
@@ -417,9 +402,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson($response);
     }
 
-    /**
-     * @Route("/set-appearance", name="opendxp_admin_settings_appearance_set", methods={"PUT"})
-     */
+    #[Route("/set-appearance", name: "opendxp_admin_settings_appearance_set", methods: ["PUT"])]
     public function setAppearanceSystemAction(
         Request $request,
         KernelInterface $kernel,
@@ -451,9 +434,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/set-system", name="opendxp_admin_settings_setsystem", methods={"PUT"})
-     */
+    #[Route("/set-system", name: "opendxp_admin_settings_setsystem", methods: ["PUT"])]
     public function setSystemAction(
         Request $request,
         KernelInterface $kernel,
@@ -485,9 +466,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/clear-cache", name="opendxp_admin_settings_clearcache", methods={"DELETE"})
-     */
+    #[Route("/clear-cache", name: "opendxp_admin_settings_clearcache", methods: ["DELETE"])]
     public function clearCacheAction(
         Request $request,
         KernelInterface $kernel,
@@ -596,9 +575,7 @@ class SettingsController extends AdminAbstractController
         }
     }
 
-    /**
-     * @Route("/clear-output-cache", name="opendxp_admin_settings_clearoutputcache", methods={"DELETE"})
-     */
+    #[Route("/clear-output-cache", name: "opendxp_admin_settings_clearoutputcache", methods: ["DELETE"])]
     public function clearOutputCacheAction(EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         $this->checkPermission('clear_fullpage_cache');
@@ -614,9 +591,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/clear-temporary-files", name="opendxp_admin_settings_cleartemporaryfiles", methods={"DELETE"})
-     */
+    #[Route("/clear-temporary-files", name: "opendxp_admin_settings_cleartemporaryfiles", methods: ["DELETE"])]
     public function clearTemporaryFilesAction(EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         $this->checkPermission('clear_temp_files');
@@ -635,9 +610,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/get-available-admin-languages", name="opendxp_admin_settings_getavailableadminlanguages", methods={"GET"})
-     */
+    #[Route("/get-available-admin-languages", name: "opendxp_admin_settings_getavailableadminlanguages", methods: ["GET"])]
     public function getAvailableAdminLanguagesAction(Request $request): JsonResponse
     {
         $langs = [];
@@ -660,9 +633,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson($langs);
     }
 
-    /**
-     * @Route("/get-available-sites", name="opendxp_admin_settings_getavailablesites", methods={"GET"})
-     */
+    #[Route("/get-available-sites", name: "opendxp_admin_settings_getavailablesites", methods: ["GET"])]
     public function getAvailableSitesAction(Request $request): JsonResponse
     {
         try {
@@ -710,9 +681,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson($sites);
     }
 
-    /**
-     * @Route("/get-available-countries", name="opendxp_admin_settings_getavailablecountries", methods={"GET"})
-     */
+    #[Route("/get-available-countries", name: "opendxp_admin_settings_getavailablecountries", methods: ["GET"])]
     public function getAvailableCountriesAction(LocaleServiceInterface $localeService): JsonResponse
     {
         $countries = $localeService->getDisplayRegions();
@@ -734,9 +703,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson($result);
     }
 
-    /**
-     * @Route("/thumbnail-adapter-check", name="opendxp_admin_settings_thumbnailadaptercheck", methods={"GET"})
-     */
+    #[Route("/thumbnail-adapter-check", name: "opendxp_admin_settings_thumbnailadaptercheck", methods: ["GET"])]
     public function thumbnailAdapterCheckAction(Request $request, TranslatorInterface $translator): Response
     {
         $content = '';
@@ -751,9 +718,7 @@ class SettingsController extends AdminAbstractController
         return new Response($content);
     }
 
-    /**
-     * @Route("/thumbnail-tree", name="opendxp_admin_settings_thumbnailtree", methods={"GET", "POST"})
-     */
+    #[Route("/thumbnail-tree", name: "opendxp_admin_settings_thumbnailtree", methods: ["GET", "POST"])]
     public function thumbnailTreeAction(): JsonResponse
     {
         $this->checkPermission('thumbnails');
@@ -805,9 +770,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson($thumbnails);
     }
 
-    /**
-     * @Route("/thumbnail-downloadable", name="opendxp_admin_settings_thumbnaildownloadable", methods={"GET"})
-     */
+    #[Route("/thumbnail-downloadable", name: "opendxp_admin_settings_thumbnaildownloadable", methods: ["GET"])]
     public function thumbnailDownloadableAction(): JsonResponse
     {
         $thumbnails = [];
@@ -827,9 +790,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson($thumbnails);
     }
 
-    /**
-     * @Route("/thumbnail-add", name="opendxp_admin_settings_thumbnailadd", methods={"POST"})
-     */
+    #[Route("/thumbnail-add", name: "opendxp_admin_settings_thumbnailadd", methods: ["POST"])]
     public function thumbnailAddAction(Request $request): JsonResponse
     {
         $this->checkPermission('thumbnails');
@@ -857,9 +818,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson(['success' => $success, 'id' => $pipe->getName()]);
     }
 
-    /**
-     * @Route("/thumbnail-delete", name="opendxp_admin_settings_thumbnaildelete", methods={"DELETE"})
-     */
+    #[Route("/thumbnail-delete", name: "opendxp_admin_settings_thumbnaildelete", methods: ["DELETE"])]
     public function thumbnailDeleteAction(Request $request): JsonResponse
     {
         $this->checkPermission('thumbnails');
@@ -875,9 +834,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/thumbnail-get", name="opendxp_admin_settings_thumbnailget", methods={"GET"})
-     */
+    #[Route("/thumbnail-get", name: "opendxp_admin_settings_thumbnailget", methods: ["GET"])]
     public function thumbnailGetAction(Request $request): JsonResponse
     {
         $this->checkPermission('thumbnails');
@@ -889,9 +846,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson($data);
     }
 
-    /**
-     * @Route("/thumbnail-update", name="opendxp_admin_settings_thumbnailupdate", methods={"PUT"})
-     */
+    #[Route("/thumbnail-update", name: "opendxp_admin_settings_thumbnailupdate", methods: ["PUT"])]
     public function thumbnailUpdateAction(Request $request): JsonResponse
     {
         $this->checkPermission('thumbnails');
@@ -941,9 +896,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/video-thumbnail-adapter-check", name="opendxp_admin_settings_videothumbnailadaptercheck", methods={"GET"})
-     */
+    #[Route("/video-thumbnail-adapter-check", name: "opendxp_admin_settings_videothumbnailadaptercheck", methods: ["GET"])]
     public function videoThumbnailAdapterCheckAction(Request $request, TranslatorInterface $translator): Response
     {
         $content = '';
@@ -957,9 +910,7 @@ class SettingsController extends AdminAbstractController
         return new Response($content);
     }
 
-    /**
-     * @Route("/video-thumbnail-tree", name="opendxp_admin_settings_videothumbnailtree", methods={"GET", "POST"})
-     */
+    #[Route("/video-thumbnail-tree", name: "opendxp_admin_settings_videothumbnailtree", methods: ["GET", "POST"])]
     public function videoThumbnailTreeAction(): JsonResponse
     {
         $this->checkPermission('thumbnails');
@@ -1011,9 +962,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson($thumbnails);
     }
 
-    /**
-     * @Route("/video-thumbnail-list", name="opendxp_admin_settings_videothumbnail_list", methods={"GET"})
-     */
+    #[Route("/video-thumbnail-list", name: "opendxp_admin_settings_videothumbnail_list", methods: ["GET"])]
     public function videoThumbnailListAction(): JsonResponse
     {
         $thumbnails = [
@@ -1031,9 +980,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson($thumbnails);
     }
 
-    /**
-     * @Route("/video-thumbnail-add", name="opendxp_admin_settings_videothumbnailadd", methods={"POST"})
-     */
+    #[Route("/video-thumbnail-add", name: "opendxp_admin_settings_videothumbnailadd", methods: ["POST"])]
     public function videoThumbnailAddAction(Request $request): JsonResponse
     {
         $this->checkPermission('thumbnails');
@@ -1061,9 +1008,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson(['success' => $success, 'id' => $pipe->getName()]);
     }
 
-    /**
-     * @Route("/video-thumbnail-delete", name="opendxp_admin_settings_videothumbnaildelete", methods={"DELETE"})
-     */
+    #[Route("/video-thumbnail-delete", name: "opendxp_admin_settings_videothumbnaildelete", methods: ["DELETE"])]
     public function videoThumbnailDeleteAction(Request $request): JsonResponse
     {
         $this->checkPermission('thumbnails');
@@ -1079,9 +1024,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/video-thumbnail-get", name="opendxp_admin_settings_videothumbnailget", methods={"GET"})
-     */
+    #[Route("/video-thumbnail-get", name: "opendxp_admin_settings_videothumbnailget", methods: ["GET"])]
     public function videoThumbnailGetAction(Request $request): JsonResponse
     {
         $this->checkPermission('thumbnails');
@@ -1094,9 +1037,7 @@ class SettingsController extends AdminAbstractController
         return $this->adminJson($data);
     }
 
-    /**
-     * @Route("/video-thumbnail-update", name="opendxp_admin_settings_videothumbnailupdate", methods={"PUT"})
-     */
+    #[Route("/video-thumbnail-update", name: "opendxp_admin_settings_videothumbnailupdate", methods: ["PUT"])]
     public function videoThumbnailUpdateAction(Request $request): JsonResponse
     {
         $this->checkPermission('thumbnails');
@@ -1143,10 +1084,9 @@ class SettingsController extends AdminAbstractController
     }
 
     /**
-     * @Route("/website-settings", name="opendxp_admin_settings_websitesettings", methods={"POST"})
-     *
      * @throws \Exception
      */
+    #[Route("/website-settings", name: "opendxp_admin_settings_websitesettings", methods: ["POST"])]
     public function websiteSettingsAction(Request $request): JsonResponse
     {
         $this->checkPermission('website_settings');
@@ -1273,9 +1213,7 @@ class SettingsController extends AdminAbstractController
         return $resultItem;
     }
 
-    /**
-     * @Route("/get-available-algorithms", name="opendxp_admin_settings_getavailablealgorithms", methods={"GET"})
-     */
+    #[Route("/get-available-algorithms", name: "opendxp_admin_settings_getavailablealgorithms", methods: ["GET"])]
     public function getAvailableAlgorithmsAction(Request $request): JsonResponse
     {
         $options = [
