@@ -22,19 +22,16 @@ use OpenDxp\Model\Element\Tag;
 use Symfony\Component\EventDispatcher\GenericEvent;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
 /**
- * @Route("/tags")
- *
  * @internal
  */
+#[Route("/tags")]
 class TagsController extends AdminAbstractController
 {
-    /**
-     * @Route("/add", name="opendxp_admin_tags_add", methods={"POST"})
-     */
+    #[Route("/add", name: "opendxp_admin_tags_add", methods: ["POST"])]
     public function addAction(Request $request): JsonResponse
     {
         $this->checkPermission('tags_configuration');
@@ -52,10 +49,9 @@ class TagsController extends AdminAbstractController
     }
 
     /**
-     * @Route("/delete", name="opendxp_admin_tags_delete", methods={"DELETE"})
-     *
      * @throws \Exception
      */
+    #[Route("/delete", name: "opendxp_admin_tags_delete", methods: ["DELETE"])]
     public function deleteAction(Request $request): JsonResponse
     {
         $this->checkPermission('tags_configuration');
@@ -71,10 +67,9 @@ class TagsController extends AdminAbstractController
     }
 
     /**
-     * @Route("/update", name="opendxp_admin_tags_update", methods={"PUT"})
-     *
      * @throws \Exception
      */
+    #[Route("/update", name: "opendxp_admin_tags_update", methods: ["PUT"])]
     public function updateAction(Request $request): JsonResponse
     {
         $this->checkPermission('tags_configuration');
@@ -97,9 +92,7 @@ class TagsController extends AdminAbstractController
         }
     }
 
-    /**
-     * @Route("/tree-get-children-by-id", name="opendxp_admin_tags_treegetchildrenbyid", methods={"GET"})
-     */
+    #[Route("/tree-get-children-by-id", name: "opendxp_admin_tags_treegetchildrenbyid", methods: ["GET"])]
     public function treeGetChildrenByIdAction(Request $request): JsonResponse
     {
         $showSelection = $request->get('showSelection') == 'true';
@@ -183,9 +176,7 @@ class TagsController extends AdminAbstractController
         return $tagArray;
     }
 
-    /**
-     * @Route("/load-tags-for-element", name="opendxp_admin_tags_loadtagsforelement", methods={"GET"})
-     */
+    #[Route("/load-tags-for-element", name: "opendxp_admin_tags_loadtagsforelement", methods: ["GET"])]
     public function loadTagsForElementAction(Request $request): JsonResponse
     {
         $assginmentCId = (int)$request->get('assignmentCId');
@@ -203,9 +194,7 @@ class TagsController extends AdminAbstractController
         return $this->adminJson($assignedTagArray);
     }
 
-    /**
-     * @Route("/add-tag-to-element", name="opendxp_admin_tags_addtagtoelement", methods={"PUT"})
-     */
+    #[Route("/add-tag-to-element", name: "opendxp_admin_tags_addtagtoelement", methods: ["PUT"])]
     public function addTagToElementAction(Request $request): JsonResponse
     {
         $assginmentCId = (int)$request->get('assignmentElementId');
@@ -222,9 +211,7 @@ class TagsController extends AdminAbstractController
         }
     }
 
-    /**
-     * @Route("/remove-tag-from-element", name="opendxp_admin_tags_removetagfromelement", methods={"DELETE"})
-     */
+    #[Route("/remove-tag-from-element", name: "opendxp_admin_tags_removetagfromelement", methods: ["DELETE"])]
     public function removeTagFromElementAction(Request $request): JsonResponse
     {
         $assginmentCId = (int)$request->get('assignmentElementId');
@@ -241,9 +228,7 @@ class TagsController extends AdminAbstractController
         }
     }
 
-    /**
-     * @Route("/get-batch-assignment-jobs", name="opendxp_admin_tags_getbatchassignmentjobs", methods={"GET"})
-     */
+    #[Route("/get-batch-assignment-jobs", name: "opendxp_admin_tags_getbatchassignmentjobs", methods: ["GET"])]
     public function getBatchAssignmentJobsAction(Request $request, EventDispatcherInterface $eventDispatcher): JsonResponse
     {
         $elementId = (int)$request->get('elementId');
@@ -375,9 +360,7 @@ class TagsController extends AdminAbstractController
         return $childrenList->loadIdList();
     }
 
-    /**
-     * @Route("/do-batch-assignment", name="opendxp_admin_tags_dobatchassignment", methods={"PUT"})
-     */
+    #[Route("/do-batch-assignment", name: "opendxp_admin_tags_dobatchassignment", methods: ["PUT"])]
     public function doBatchAssignmentAction(Request $request): JsonResponse
     {
         $cType = strip_tags($request->get('elementType', ''));
