@@ -25,6 +25,7 @@ use OpenDxp\Model\Asset;
 use OpenDxp\Model\DataObject;
 use OpenDxp\Model\Element;
 use OpenDxp\Model\User;
+use OpenDxp\Model\User\Workspace;
 use OpenDxp\SystemSettingsConfig;
 use OpenDxp\Tool;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
@@ -34,7 +35,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Session\Attribute\AttributeBagInterface;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 use Symfony\Component\HttpKernel\Event\ControllerEvent;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Routing\RouterInterface;
 use Symfony\Component\Security\Core\Validator\Constraints\UserPassword;
@@ -46,9 +47,7 @@ use Symfony\Contracts\Translation\TranslatorInterface;
  */
 class UserController extends AdminAbstractController implements KernelControllerEventInterface
 {
-    /**
-     * @Route("/user/tree-get-children-by-id", name="opendxp_admin_user_treegetchildrenbyid", methods={"GET"})
-     */
+    #[Route("/user/tree-get-children-by-id", name: "opendxp_admin_user_treegetchildrenbyid", methods: ["GET"])]
     public function treeGetChildrenByIdAction(Request $request): JsonResponse
     {
         $list = new User\Listing();
@@ -106,9 +105,7 @@ class UserController extends AdminAbstractController implements KernelController
         return $tmpUser;
     }
 
-    /**
-     * @Route("/user/add", name="opendxp_admin_user_add", methods={"POST"})
-     */
+    #[Route("/user/add", name: "opendxp_admin_user_add", methods: ["POST"])]
     public function addAction(Request $request): JsonResponse
     {
         try {
@@ -224,10 +221,9 @@ class UserController extends AdminAbstractController implements KernelController
     }
 
     /**
-     * @Route("/user/delete", name="opendxp_admin_user_delete", methods={"DELETE"})
-     *
      * @throws \Exception
      */
+    #[Route("/user/delete", name: "opendxp_admin_user_delete", methods: ["DELETE"])]
     public function deleteAction(Request $request): JsonResponse
     {
         $user = User\AbstractUser::getById((int)$request->get('id'));
@@ -257,10 +253,9 @@ class UserController extends AdminAbstractController implements KernelController
     }
 
     /**
-     * @Route("/user/update", name="opendxp_admin_user_update", methods={"PUT"})
-     *
      * @throws \Exception
      */
+    #[Route("/user/update", name: "opendxp_admin_user_update", methods: ["PUT"])]
     public function updateAction(Request $request, TranslatorInterface $translator): JsonResponse
     {
         /** @var User|User\Role|null $user */
@@ -366,10 +361,9 @@ class UserController extends AdminAbstractController implements KernelController
     }
 
     /**
-     * @Route("/user/get", name="opendxp_admin_user_get", methods={"GET"})
-     *
      * @throws \Exception
      */
+    #[Route("/user/get", name: "opendxp_admin_user_get", methods: ["GET"])]
     public function getAction(Request $request): JsonResponse
     {
         $userId = (int)$request->get('id');
@@ -390,6 +384,7 @@ class UserController extends AdminAbstractController implements KernelController
         // workspaces
         $types = ['asset', 'document', 'object'];
         foreach ($types as $type) {
+            /** @var Workspace\Document[]|Workspace\Asset[]|Workspace\DataObject[] $workspaces */
             $workspaces = $user->{'getWorkspaces' . ucfirst($type)}();
             foreach ($workspaces as $wKey => $workspace) {
                 $el = Element\Service::getElementById($type, $workspace->getCid());
@@ -472,9 +467,7 @@ class UserController extends AdminAbstractController implements KernelController
         ]);
     }
 
-    /**
-     * @Route("/user/get-minimal", name="opendxp_admin_user_getminimal", methods={"GET"})
-     */
+    #[Route("/user/get-minimal", name: "opendxp_admin_user_getminimal", methods: ["GET"])]
     public function getMinimalAction(Request $request): JsonResponse
     {
         $user = User::getById((int)$request->get('id'));
@@ -493,9 +486,7 @@ class UserController extends AdminAbstractController implements KernelController
         return $this->adminJson($minimalUserData);
     }
 
-    /**
-     * @Route("/user/upload-current-user-image", name="opendxp_admin_user_uploadcurrentuserimage", methods={"POST"})
-     */
+    #[Route("/user/upload-current-user-image", name: "opendxp_admin_user_uploadcurrentuserimage", methods: ["POST"])]
     public function uploadCurrentUserImageAction(Request $request): JsonResponse
     {
         $user = $this->getAdminUser();
@@ -512,9 +503,7 @@ class UserController extends AdminAbstractController implements KernelController
         }
     }
 
-    /**
-     * @Route("/user/update-current-user", name="opendxp_admin_user_updatecurrentuser", methods={"PUT"})
-     */
+    #[Route("/user/update-current-user", name: "opendxp_admin_user_updatecurrentuser", methods: ["PUT"])]
     public function updateCurrentUserAction(Request $request, ValidatorInterface $validator): JsonResponse
     {
         //TODO Can be completely validated with Symfony Validator
@@ -600,9 +589,7 @@ class UserController extends AdminAbstractController implements KernelController
         }
     }
 
-    /**
-     * @Route("/user/get-current-user", name="opendxp_admin_user_getcurrentuser", methods={"GET"})
-     */
+    #[Route("/user/get-current-user", name: "opendxp_admin_user_getcurrentuser", methods: ["GET"])]
     public function getCurrentUserAction(Request $request): Response
     {
         $user = $this->getAdminUser();
@@ -640,9 +627,7 @@ class UserController extends AdminAbstractController implements KernelController
 
     // ROLES
 
-    /**
-     * @Route("/user/role-tree-get-children-by-id", name="opendxp_admin_user_roletreegetchildrenbyid", methods={"GET"})
-     */
+    #[Route("/user/role-tree-get-children-by-id", name: "opendxp_admin_user_roletreegetchildrenbyid", methods: ["GET"])]
     public function roleTreeGetChildrenByIdAction(Request $request): JsonResponse
     {
         $list = new User\Role\Listing();
@@ -691,9 +676,7 @@ class UserController extends AdminAbstractController implements KernelController
         return $tmpUser;
     }
 
-    /**
-     * @Route("/user/role-get", name="opendxp_admin_user_roleget", methods={"GET"})
-     */
+    #[Route("/user/role-get", name: "opendxp_admin_user_roleget", methods: ["GET"])]
     public function roleGetAction(Request $request): JsonResponse
     {
         $role = User\Role::getById((int)$request->get('id'));
@@ -705,6 +688,7 @@ class UserController extends AdminAbstractController implements KernelController
         // workspaces
         $types = ['asset', 'document', 'object'];
         foreach ($types as $type) {
+            /** @var Workspace\Document[]|Workspace\Asset[]|Workspace\DataObject[] $workspaces */
             $workspaces = $role->{'getWorkspaces' . ucfirst($type)}();
             foreach ($workspaces as $wKey => $workspace) {
                 $el = Element\Service::getElementById($type, $workspace->getCid());
@@ -742,10 +726,9 @@ class UserController extends AdminAbstractController implements KernelController
     }
 
     /**
-     * @Route("/user/upload-image", name="opendxp_admin_user_uploadimage", methods={"POST"})
-     *
      * @throws \Exception
      */
+    #[Route("/user/upload-image", name: "opendxp_admin_user_uploadimage", methods: ["POST"])]
     public function uploadImageAction(Request $request): JsonResponse
     {
         $userObj = User::getById($this->getUserId($request));
@@ -779,10 +762,9 @@ class UserController extends AdminAbstractController implements KernelController
     }
 
     /**
-     * @Route("/user/delete-image", name="opendxp_admin_user_deleteimage", methods={"DELETE"})
-     *
      * @throws \Exception
      */
+    #[Route("/user/delete-image", name: "opendxp_admin_user_deleteimage", methods: ["DELETE"])]
     public function deleteImageAction(Request $request): JsonResponse
     {
         $userObj = User::getById($this->getUserId($request));
@@ -808,9 +790,7 @@ class UserController extends AdminAbstractController implements KernelController
         return $this->adminJson(['success' => true]);
     }
 
-    /**
-     * @Route("/user/disable-2fa", name="opendxp_admin_user_disable2fasecret", methods={"DELETE"})
-     */
+    #[Route("/user/disable-2fa", name: "opendxp_admin_user_disable2fasecret", methods: ["DELETE"])]
     public function disable2FaSecretAction(Request $request): JsonResponse
     {
         $user = $this->getAdminUser();
@@ -828,9 +808,7 @@ class UserController extends AdminAbstractController implements KernelController
         ]);
     }
 
-    /**
-     * @Route("/user/reset-2fa-secret", name="opendxp_admin_user_reset2fasecret", methods={"PUT"})
-     */
+    #[Route("/user/reset-2fa-secret", name: "opendxp_admin_user_reset2fasecret", methods: ["PUT"])]
     public function reset2FaSecretAction(Request $request): JsonResponse
     {
         $user = User::getById((int)$request->get('id'));
@@ -846,9 +824,7 @@ class UserController extends AdminAbstractController implements KernelController
         ]);
     }
 
-    /**
-     * @Route("/user/reset-my-2fa-secret", name="opendxp_admin_user_reset_my_2fa_secret", methods={"PUT"})
-     */
+    #[Route("/user/reset-my-2fa-secret", name: "opendxp_admin_user_reset_my_2fa_secret", methods: ["PUT"])]
     public function resetMy2FaSecretAction(Request $request): JsonResponse
     {
         $user = $this->getAdminUser();
@@ -862,9 +838,7 @@ class UserController extends AdminAbstractController implements KernelController
         ]);
     }
 
-    /**
-     * @Route("/user/get-image", name="opendxp_admin_user_getimage", methods={"GET"})
-     */
+    #[Route("/user/get-image", name: "opendxp_admin_user_getimage", methods: ["GET"])]
     public function getImageAction(Request $request): StreamedResponse
     {
         $userObj = User::getById($this->getUserId($request));
@@ -881,10 +855,9 @@ class UserController extends AdminAbstractController implements KernelController
     }
 
     /**
-     * @Route("/user/get-token-login-link", name="opendxp_admin_user_gettokenloginlink", methods={"GET"})
-     *
      * @throws \Exception
      */
+    #[Route("/user/get-token-login-link", name: "opendxp_admin_user_gettokenloginlink", methods: ["GET"])]
     public function getTokenLoginLinkAction(Request $request, TranslatorInterface $translator): JsonResponse
     {
         $user = User::getById((int) $request->get('id'));
@@ -921,9 +894,7 @@ class UserController extends AdminAbstractController implements KernelController
         ]);
     }
 
-    /**
-     * @Route("/user/search", name="opendxp_admin_user_search", methods={"GET"})
-     */
+    #[Route("/user/search", name: "opendxp_admin_user_search", methods: ["GET"])]
     public function searchAction(Request $request): JsonResponse
     {
         $q = '%' . $request->get('query') . '%';
@@ -971,9 +942,7 @@ class UserController extends AdminAbstractController implements KernelController
         $this->checkActionPermission($event, 'users', $unrestrictedActions);
     }
 
-    /**
-     * @Route("/user/get-users-for-sharing", name="opendxp_admin_user_getusersforsharing", methods={"GET"})
-     */
+    #[Route("/user/get-users-for-sharing", name: "opendxp_admin_user_getusersforsharing", methods: ["GET"])]
     public function getUsersForSharingAction(Request $request): JsonResponse
     {
         $this->checkPermission('share_configurations');
@@ -981,9 +950,7 @@ class UserController extends AdminAbstractController implements KernelController
         return $this->getUsersAction($request);
     }
 
-    /**
-     * @Route("/user/get-roles-for-sharing", name="opendxp_admin_user_getrolesforsharing", methods={"GET"}))
-     */
+    #[Route("/user/get-roles-for-sharing", name: "opendxp_admin_user_getrolesforsharing", methods: ["GET"])]
     public function getRolesForSharingAction(Request $request): JsonResponse
     {
         $this->checkPermission('share_configurations');
@@ -991,9 +958,7 @@ class UserController extends AdminAbstractController implements KernelController
         return $this->getRolesAction($request);
     }
 
-    /**
-     * @Route("/user/get-users", name="opendxp_admin_user_getusers", methods={"GET"})
-     */
+    #[Route("/user/get-users", name: "opendxp_admin_user_getusers", methods: ["GET"])]
     public function getUsersAction(Request $request): JsonResponse
     {
         $users = [];
@@ -1024,9 +989,7 @@ class UserController extends AdminAbstractController implements KernelController
         return $this->adminJson(['success' => true, 'total' => count($users), 'data' => $users]);
     }
 
-    /**
-     * @Route("/user/get-roles", name="opendxp_admin_user_getroles", methods={"GET"})
-     */
+    #[Route("/user/get-roles", name: "opendxp_admin_user_getroles", methods: ["GET"])]
     public function getRolesAction(Request $request): JsonResponse
     {
         $roles = [];
@@ -1048,19 +1011,16 @@ class UserController extends AdminAbstractController implements KernelController
         return $this->adminJson(['success' => true, 'total' => count($roles), 'data' => $roles]);
     }
 
-    /**
-     * @Route("/user/get-default-key-bindings", name="opendxp_admin_user_getdefaultkeybindings", methods={"GET"})
-     */
+    #[Route("/user/get-default-key-bindings", name: "opendxp_admin_user_getdefaultkeybindings", methods: ["GET"])]
     public function getDefaultKeyBindingsAction(Request $request): JsonResponse
     {
         return $this->adminJson(['success' => true, 'data' => UserHelper::getDefaultKeyBindings()]);
     }
 
     /**
-     * @Route("/user/invitationlink", name="opendxp_admin_user_invitationlink", methods={"POST"})
-     *
      * @throws \Exception
      */
+    #[Route("/user/invitationlink", name: "opendxp_admin_user_invitationlink", methods: ["POST"])]
     public function invitationLinkAction(Request $request, TranslatorInterface $translator, RouterInterface $router): JsonResponse
     {
         $success = false;
