@@ -22,7 +22,7 @@ use OpenDxp\Bundle\AdminBundle\GDPR\DataProvider\Manager;
 use OpenDxp\Controller\KernelControllerEventInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpKernel\Event\ControllerEvent;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 
 /**
  *
@@ -30,9 +30,7 @@ use Symfony\Component\Routing\Annotation\Route;
  */
 class AdminController extends AdminAbstractController implements KernelControllerEventInterface
 {
-    /**
-     * @Route("/get-data-providers", name="opendxp_admin_gdpr_admin_getdataproviders", methods={"GET"})
-     */
+    #[Route("/get-data-providers", name: "opendxp_admin_gdpr_admin_getdataproviders", methods: ["GET"])]
     public function getDataProvidersAction(Manager $manager): JsonResponse
     {
         $response = [];
