@@ -22,15 +22,14 @@ use OpenDxp\Model\Tool\Email\Log;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Event\ControllerEvent;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 
 /**
  * Class SentMailController
  *
- * @Route("/sent-mail")
- *
  * @internal
  */
+#[Route("/sent-mail")]
 class SentMailController extends AdminAbstractController implements KernelControllerEventInterface
 {
     public function onKernelControllerEvent(ControllerEvent $event): void
@@ -42,9 +41,7 @@ class SentMailController extends AdminAbstractController implements KernelContro
         $this->checkActionPermission($event, 'gdpr_data_extractor');
     }
 
-    /**
-     * @Route("/export", name="opendxp_admin_gdpr_sentmail_exportdataobject", methods={"GET"})
-     */
+    #[Route("/export", name: "opendxp_admin_gdpr_sentmail_exportdataobject", methods: ["GET"])]
     public function exportDataObjectAction(Request $request): JsonResponse
     {
         $this->checkPermission('emails');
