@@ -27,20 +27,18 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Profiler\Profiler;
 use Symfony\Component\Mime\Address;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 
 /**
- * @Route("/email")
- *
  * @internal
  */
+#[Route("/email")]
 class EmailController extends AdminAbstractController
 {
     /**
-     * @Route("/email-logs", name="opendxp_admin_email_emaillogs", methods={"GET", "POST"})
-     *
      * @throws \Exception
      */
+    #[Route("/email-logs", name: "opendxp_admin_email_emaillogs", methods: ["GET", "POST"])]
     public function emailLogsAction(Request $request): JsonResponse
     {
         if (!$this->getAdminUser()->isAllowed('emails') && !$this->getAdminUser()->isAllowed('gdpr_data_extractor')) {
@@ -111,10 +109,9 @@ class EmailController extends AdminAbstractController
     }
 
     /**
-     * @Route("/show-email-log", name="opendxp_admin_email_showemaillog", methods={"GET"})
-     *
      * @throws \Exception
      */
+    #[Route("/show-email-log", name: "opendxp_admin_email_showemaillog", methods: ["GET"])]
     public function showEmailLogAction(Request $request, ?Profiler $profiler): JsonResponse|Response
     {
         if ($profiler) {
@@ -240,10 +237,9 @@ class EmailController extends AdminAbstractController
     }
 
     /**
-     * @Route("/delete-email-log", name="opendxp_admin_email_deleteemaillog", methods={"DELETE"})
-     *
      * @throws \Exception
      */
+    #[Route("/delete-email-log", name: "opendxp_admin_email_deleteemaillog", methods: ["DELETE"])]
     public function deleteEmailLogAction(Request $request): JsonResponse
     {
         if (!$this->getAdminUser()->isAllowed('emails')) {
@@ -263,10 +259,9 @@ class EmailController extends AdminAbstractController
     }
 
     /**
-     * @Route("/resend-email", name="opendxp_admin_email_resendemail", methods={"POST"})
-     *
      * @throws \Exception
      */
+    #[Route("/resend-email", name: "opendxp_admin_email_resendemail", methods: ["POST"])]
     public function resendEmailAction(Request $request): JsonResponse
     {
         if (!$this->getAdminUser()->isAllowed('emails')) {
@@ -354,10 +349,9 @@ class EmailController extends AdminAbstractController
     }
 
     /**
-     * @Route("/send-test-email", name="opendxp_admin_email_sendtestemail", methods={"POST"})
-     *
      * @throws \Exception
      */
+    #[Route("/send-test-email", name: "opendxp_admin_email_sendtestemail", methods: ["POST"])]
     public function sendTestEmailAction(Request $request): JsonResponse
     {
         if (!$this->getAdminUser()->isAllowed('emails')) {
@@ -418,10 +412,9 @@ class EmailController extends AdminAbstractController
     }
 
     /**
-     * @Route("/blocklist", name="opendxp_admin_email_blocklist", methods={"POST"})
-     *
      * @throws \Exception
      */
+    #[Route("/blocklist", name: "opendxp_admin_email_blocklist", methods: ["POST"])]
     public function blocklistAction(Request $request): JsonResponse
     {
         if (!$this->getAdminUser()->isAllowed('emails')) {
